refactor(editor): split EditorView.swift into focused files (#253)
Move LinkAwareTextView, its markdown-styling extension, and MarkdownTheme
out of EditorView.swift as pure relocations. Combined with the prior
checkpoint commit, EditorView.swift drops from 6,121 to 1,196 lines and
now contains only EditorView, RawEditor (+Coordinator), the pending-signal
forwarders, and small editor helpers.

New files:
- LinkAwareTextView.swift (class + debugLog helper)
- LinkAwareTextView+MarkdownStyling.swift (styling/preview engine,
  collapsible toggles, inline AI bar presentation, regex cache,
  styleMarkdownContent, custom NSAttributedString keys)
- MarkdownTheme.swift (theme + private NSFont.withWeight helper)

Access-level bumps required by the file split (single module, no
behavior change): EditorFontSignature private->internal;
aiSparkleTapped, shouldSkipIncrementalMarkdownRestyle and 21
LinkAwareTextView stored properties fileprivate/private->internal.
Trailing whitespace stripped from moved lines per repo rules.

Full suite: 2043 tests, 0 failures, 1 skipped (baseline match).
