+++ macOS/SynapseNotes/EditorView.swift
@@ -1181,147 +1181,7 @@ func refreshActiveEditorForHideMarkdownToggle(hideMarkdown: Bool) {
     refreshEditorForHideMarkdownToggle(textView, hideMarkdown: hideMarkdown)
 }
 
-// MARK: - Markdown styling theme
-
-struct MarkdownTheme {
-    // MARK: - Font functions based on SettingsManager
-    
-    static func bodyFont(for settings: SettingsManager) -> NSFont {
-        let size = CGFloat(settings.editorFontSize)
-        if settings.editorBodyFontFamily.isEmpty || settings.editorBodyFontFamily == "System" {
-            return NSFont.systemFont(ofSize: size)
-        }
-        return NSFont(name: settings.editorBodyFontFamily, size: size) ?? NSFont.systemFont(ofSize: size)
-    }
-    
-    static func monoFont(for settings: SettingsManager) -> NSFont {
-        let baseSize = CGFloat(settings.editorFontSize)
-        let size = max(10, baseSize / SynapseTheme.Layout.phi)
-        if settings.editorMonospaceFontFamily.isEmpty || settings.editorMonospaceFontFamily == "System Monospace" {
-            return NSFont.monospacedSystemFont(ofSize: size, weight: .regular)
-        }
-        return NSFont(name: settings.editorMonospaceFontFamily, size: size) 
-            ?? NSFont.monospacedSystemFont(ofSize: size, weight: .regular)
-    }
-    
-    static func h1Font(for settings: SettingsManager) -> NSFont {
-        let size = round(CGFloat(settings.editorFontSize) * SynapseTheme.Editor.headingH1Multiplier)
-        if settings.editorBodyFontFamily.isEmpty || settings.editorBodyFontFamily == "System" {
-            return NSFont.systemFont(ofSize: size, weight: .bold)
-        }
-        return NSFont(name: settings.editorBodyFontFamily, size: size)?.withWeight(.bold) 
-            ?? NSFont.systemFont(ofSize: size, weight: .bold)
-    }
-    
-    static func h2Font(for settings: SettingsManager) -> NSFont {
-        let size = round(CGFloat(settings.editorFontSize) * SynapseTheme.Editor.headingH2Multiplier)
-        if settings.editorBodyFontFamily.isEmpty || settings.editorBodyFontFamily == "System" {
-            return NSFont.systemFont(ofSize: size, weight: .bold)
-        }
-        return NSFont(name: settings.editorBodyFontFamily, size: size)?.withWeight(.bold)
-            ?? NSFont.systemFont(ofSize: size, weight: .bold)
-    }
-    
-    static func h3Font(for settings: SettingsManager) -> NSFont {
-        let size = round(CGFloat(settings.editorFontSize) * SynapseTheme.Editor.headingH3Multiplier)
-        if settings.editorBodyFontFamily.isEmpty || settings.editorBodyFontFamily == "System" {
-            return NSFont.systemFont(ofSize: size, weight: .semibold)
-        }
-        return NSFont(name: settings.editorBodyFontFamily, size: size)?.withWeight(.semibold)
-            ?? NSFont.systemFont(ofSize: size, weight: .semibold)
-    }
-    
-    static func h4Font(for settings: SettingsManager) -> NSFont {
-        let size = round(CGFloat(settings.editorFontSize) * SynapseTheme.Editor.headingH4Multiplier)
-        if settings.editorBodyFontFamily.isEmpty || settings.editorBodyFontFamily == "System" {
-            return NSFont.systemFont(ofSize: size, weight: .semibold)
-        }
-        return NSFont(name: settings.editorBodyFontFamily, size: size)?.withWeight(.semibold)
-            ?? NSFont.systemFont(ofSize: size, weight: .semibold)
-    }
-    
-    static func boldFont(for settings: SettingsManager) -> NSFont {
-        let size = CGFloat(settings.editorFontSize)
-        if settings.editorBodyFontFamily.isEmpty || settings.editorBodyFontFamily == "System" {
-            return NSFont.systemFont(ofSize: size, weight: .bold)
-        }
-        return NSFont(name: settings.editorBodyFontFamily, size: size)?.withWeight(.bold)
-            ?? NSFont.systemFont(ofSize: size, weight: .bold)
-    }
-    
-    static func italicFont(for settings: SettingsManager) -> NSFont {
-        let size = CGFloat(settings.editorFontSize)
-        if settings.editorBodyFontFamily.isEmpty || settings.editorBodyFontFamily == "System" {
-            let descriptor = NSFont.systemFont(ofSize: size).fontDescriptor.withSymbolicTraits(.italic)
-            return NSFont(descriptor: descriptor, size: size) ?? NSFont.systemFont(ofSize: size)
-        }
-        let baseFont = NSFont(name: settings.editorBodyFontFamily, size: size) ?? NSFont.systemFont(ofSize: size)
-        let descriptor = baseFont.fontDescriptor.withSymbolicTraits(.italic)
-        return NSFont(descriptor: descriptor, size: size) ?? baseFont
-    }
-
-    static func boldItalicFont(for settings: SettingsManager) -> NSFont {
-        let size = CGFloat(settings.editorFontSize)
-        let bold = boldFont(for: settings)
-        let descriptor = bold.fontDescriptor.withSymbolicTraits([.bold, .italic])
-        return NSFont(descriptor: descriptor, size: size) ?? bold
-    }
-
-    static func lineHeightMultiple(for settings: SettingsManager) -> CGFloat {
-        max(0.8, min(3.0, CGFloat(settings.editorLineHeight)))
-    }
-
-    /// Paragraph style whose line box matches CSS `line-height: multiple` — i.e. a
-    /// multiple of the FONT SIZE, not of the font's natural line height. NSTextView's
-    /// natural line height already bakes in the font's intrinsic leading (~1.18× for
-    /// SF), so multiplying that by the user's multiple over-spaced lines versus the
-    /// HTML preview. We set the line box directly to `fontSize * multiple`, floored at
-    /// the natural height so small multiples never clip glyphs (CSS overlaps instead of
-    /// cropping; a hard maximumLineHeight below natural would crop).
-    static func paragraphStyle(font: NSFont, lineHeightMultiple multiple: CGFloat) -> NSMutableParagraphStyle {
-        let naturalLineHeight = font.ascender - font.descender + font.leading
-        let targetLineHeight = font.pointSize * multiple
-        let style = NSMutableParagraphStyle()
-        style.minimumLineHeight = targetLineHeight
-        style.maximumLineHeight = max(targetLineHeight, naturalLineHeight)
-        style.lineSpacing = 0
-        return style
-    }
-
-    // MARK: - Legacy static constants (for backward compatibility)
-    
-    static let body = NSFont.systemFont(ofSize: 15)
-    static let mono = NSFont.monospacedSystemFont(ofSize: 13, weight: .regular)
-    static let h1   = NSFont.systemFont(ofSize: round(SynapseTheme.Editor.h1FontSize), weight: .bold)
-    static let h2   = NSFont.systemFont(ofSize: round(SynapseTheme.Editor.h2FontSize), weight: .bold)
-    static let h3   = NSFont.systemFont(ofSize: round(SynapseTheme.Editor.h3FontSize), weight: .semibold)
-    static let h4   = NSFont.systemFont(ofSize: round(SynapseTheme.Editor.h4FontSize), weight: .semibold)
-    // Use static var so these read from ThemeEnvironment.shared at each call-site
-    // rather than being frozen at class-load time.
-    static var dimColor:            NSColor { SynapseTheme.editorMuted }
-    static var tagColor:            NSColor { SynapseTheme.editorLink }
-    static var linkColor:           NSColor { SynapseTheme.editorLink }
-    static var unresolvedLinkColor: NSColor { SynapseTheme.editorUnresolvedLink }
-    static var codeBackground:      NSColor { SynapseTheme.editorCodeBackground }
-}
-
-// Helper extension to apply font weight
-private extension NSFont {
-    func withWeight(_ weight: NSFont.Weight) -> NSFont {
-        // Create a new font descriptor with the desired weight trait
-        var traits = fontDescriptor.symbolicTraits
-        // Map NSFont.Weight to NSFontDescriptor.SymbolicTraits
-        if weight == .bold || weight.rawValue >= NSFont.Weight.bold.rawValue {
-            traits.insert(.bold)
-        } else if weight == .semibold || weight.rawValue >= NSFont.Weight.semibold.rawValue {
-            traits.insert(.bold)
-        }
-        let descriptor = fontDescriptor.withSymbolicTraits(traits)
-        return NSFont(descriptor: descriptor, size: pointSize) ?? self
-    }
-}
-
-private struct EditorFontSignature: Equatable {
+struct EditorFontSignature: Equatable {
     let bodyFontFamily: String
     let monospaceFontFamily: String
     let fontSize: Int
@@ -1332,2944 +1192,5 @@ private struct EditorFontSignature: Equatable {
         monospaceFontFamily = settings?.editorMonospaceFontFamily ?? "System Monospace"
         fontSize = settings?.editorFontSize ?? 15
         lineHeight = settings?.editorLineHeight ?? 1.6
-    }
-}
-
-/// Custom attribute key for wiki links — avoids NSTextView overriding our foreground color via linkTextAttributes.
-extension NSAttributedString.Key {
-    static let wikilinkTarget = NSAttributedString.Key("Synapse.wikilinkTarget")
-    static let tagTarget = NSAttributedString.Key("Synapse.tagTarget")
-    /// Marks a character range so `LinkAwareTextView.drawBackground(in:)` draws
-    /// its background color across the full container width, not just the glyph bounds.
-    /// The value must be an `NSColor`.
-    static let codeBlockFullWidthBackground = NSAttributedString.Key("Synapse.codeBlockFullWidthBackground")
-    /// Marks a character range as belonging to a blockquote so
-    /// `LinkAwareTextView.drawBackground(in:)` can paint a decorative accent bar
-    /// along the leading edge of every line in the range. Value must be an `NSColor`.
-    static let blockquoteLeftBorder = NSAttributedString.Key("Synapse.blockquoteLeftBorder")
-}
-
-/// Thread-safe regex cache for markdown styling outside of LinkAwareTextView.
-private var sharedRegexCache: [String: NSRegularExpression] = [:]
-
-private func cachedRegex(_ pattern: String, options: NSRegularExpression.Options = []) -> NSRegularExpression? {
-    let key = "\(pattern)|\(options.rawValue)"
-    if let cached = sharedRegexCache[key] { return cached }
-    guard let compiled = try? NSRegularExpression(pattern: pattern, options: options) else { return nil }
-    sharedRegexCache[key] = compiled
-    return compiled
-}
-
-/// Styles markdown text and returns an attributed string for display
-func styleMarkdownContent(_ content: String, fontSize: CGFloat = 12) -> NSAttributedString {
-    let storage = NSTextStorage(string: content)
-    let text = content as NSString
-    let fullRange = NSRange(location: 0, length: text.length)
-
-    let baseFont = NSFont.systemFont(ofSize: fontSize)
-    storage.addAttributes([
-        .font: baseFont,
-        .foregroundColor: SynapseTheme.editorForeground,
-    ], range: fullRange)
-
-    func applyPattern(_ pattern: String, options: NSRegularExpression.Options = [], apply: (NSRange) -> Void) {
-        guard let regex = cachedRegex(pattern, options: options) else { return }
-        regex.enumerateMatches(in: content, options: [], range: fullRange) { match, _, _ in
-            guard let range = match?.range else { return }
-            apply(range)
-        }
-    }
-
-    func dimDelims(_ range: NSRange, _ delimLen: Int) {
-        guard range.length >= delimLen * 2 else { return }
-        storage.addAttribute(.foregroundColor, value: MarkdownTheme.dimColor, range: NSRange(location: range.location, length: delimLen))
-        storage.addAttribute(.foregroundColor, value: MarkdownTheme.dimColor, range: NSRange(location: range.location + range.length - delimLen, length: delimLen))
-    }
-
-    // Headers
-    let headerPatterns: [(String, NSFont)] = [
-        ("^#{6} .+$", NSFont.systemFont(ofSize: fontSize + 2, weight: .semibold)),
-        ("^#{5} .+$", NSFont.systemFont(ofSize: fontSize + 2, weight: .semibold)),
-        ("^#{4} .+$", NSFont.systemFont(ofSize: fontSize + 2, weight: .semibold)),
-        ("^### .+$",  NSFont.systemFont(ofSize: fontSize + 4, weight: .bold)),
-        ("^## .+$",   NSFont.systemFont(ofSize: fontSize + 6, weight: .bold)),
-        ("^# .+$",    NSFont.systemFont(ofSize: fontSize + 8, weight: .bold)),
-    ]
-    for (pattern, font) in headerPatterns {
-        applyPattern(pattern, options: [.anchorsMatchLines]) { range in
-            storage.addAttributes([.font: font], range: range)
-            let hashEnd = (text.substring(with: range) as NSString).range(of: "^#{1,6} ", options: .regularExpression)
-            if hashEnd.location != NSNotFound {
-                storage.addAttribute(.foregroundColor, value: MarkdownTheme.dimColor, range: NSRange(location: range.location + hashEnd.location, length: hashEnd.length))
-            }
-        }
-    }
-
-    // Italic — applied first so bold applied afterward wins on **word** spans
-    applyPattern("(?<!\\*)\\*(?!\\*)(.+?)(?<!\\*)\\*(?!\\*)") { range in
-        let desc = baseFont.fontDescriptor.withSymbolicTraits(.italic)
-        if let f = NSFont(descriptor: desc, size: fontSize) {
-            storage.addAttribute(.font, value: f, range: range)
-        }
-        dimDelims(range, 1)
-    }
-    // Bold — applied after italic so it wins over any italic applied to ** delimiters
-    applyPattern("\\*\\*(.+?)\\*\\*") { range in
-        storage.addAttribute(.font, value: NSFont.systemFont(ofSize: fontSize, weight: .bold), range: range)
-        dimDelims(range, 2)
-    }
-    applyPattern("(?<![\\w_])_(?!_)(.+?)(?<!_)_(?![\\w_])") { range in
-        let desc = baseFont.fontDescriptor.withSymbolicTraits(.italic)
-        if let f = NSFont(descriptor: desc, size: fontSize) {
-            storage.addAttribute(.font, value: f, range: range)
-        }
-        dimDelims(range, 1)
-    }
-    applyPattern("__(.+?)__") { range in
-        storage.addAttribute(.font, value: NSFont.systemFont(ofSize: fontSize, weight: .bold), range: range)
-        dimDelims(range, 2)
-    }
-    // Bold+italic — applied last so it wins over the bold and italic passes above on
-    // ***word*** spans (which would otherwise collapse to plain bold).
-    applyPattern("\\*\\*\\*(.+?)\\*\\*\\*") { range in
-        let desc = NSFont.systemFont(ofSize: fontSize, weight: .bold).fontDescriptor.withSymbolicTraits([.bold, .italic])
-        let font = NSFont(descriptor: desc, size: fontSize) ?? NSFont.systemFont(ofSize: fontSize, weight: .bold)
-        storage.addAttribute(.font, value: font, range: range)
-        dimDelims(range, 3)
-    }
-    // Strikethrough — double first, then single with word-boundary guards.
-    applyPattern("~~(.+?)~~") { range in
-        storage.addAttribute(.strikethroughStyle, value: NSUnderlineStyle.single.rawValue, range: range)
-        dimDelims(range, 2)
-    }
-    applyPattern("(?<![\\w~])~(?!~)(.+?)(?<!~)~(?![\\w~])") { range in
-        storage.addAttribute(.strikethroughStyle, value: NSUnderlineStyle.single.rawValue, range: range)
-        dimDelims(range, 1)
-    }
-    // Inline code
-    applyPattern("`([^`\\n]+)`") { range in
-        storage.addAttributes([.font: NSFont.monospacedSystemFont(ofSize: max(10, fontSize - 1), weight: .regular), .backgroundColor: MarkdownTheme.codeBackground], range: range)
-    }
-    // Code blocks
-    applyPattern("```[\\s\\S]*?```") { range in
-        storage.addAttributes([.font: NSFont.monospacedSystemFont(ofSize: max(10, fontSize - 1), weight: .regular), .backgroundColor: MarkdownTheme.codeBackground, .foregroundColor: SynapseTheme.editorForeground], range: range)
-    }
-    // Blockquotes
-    applyPattern("^> .+$", options: [.anchorsMatchLines]) { range in
-        storage.addAttribute(.foregroundColor, value: MarkdownTheme.dimColor, range: range)
-    }
-    // Inline tags
-    AppState.inlineTagMatches(in: content).forEach { match in
-        storage.addAttribute(.foregroundColor, value: MarkdownTheme.tagColor, range: match.range)
-        storage.addAttribute(.tagTarget, value: match.normalized, range: match.range)
-    }
-    // Wiki links
-    applyPattern("\\[\\[[^\\]]+\\]\\]") { range in
-        guard range.length > 4 else { return }
-        let inner = text.substring(with: NSRange(location: range.location + 2, length: range.length - 4))
-        storage.addAttributes([.foregroundColor: MarkdownTheme.linkColor, .underlineStyle: NSUnderlineStyle.single.rawValue, .link: inner], range: range)
-    }
-    // Markdown links
-    applyPattern("(?<!!)\\[([^\\]]+)\\]\\(([^)]+)\\)") { range in
-        // Need to re-match to get capture groups
-        guard let regex = cachedRegex("(?<!!)\\[([^\\]]+)\\]\\(([^)]+)\\)") else { return }
-        regex.enumerateMatches(in: content, options: [], range: range) { match, _, _ in
-            guard let match, match.numberOfRanges >= 3 else { return }
-            let full = match.range(at: 0)
-            let label = match.range(at: 1)
-            storage.addAttribute(.foregroundColor, value: MarkdownTheme.dimColor, range: full)
-            storage.addAttributes([.foregroundColor: MarkdownTheme.linkColor, .underlineStyle: NSUnderlineStyle.single.rawValue], range: label)
-        }
-    }
-    // Horizontal rules
-    applyPattern("^---$", options: [.anchorsMatchLines]) { range in
-        storage.addAttribute(.foregroundColor, value: MarkdownTheme.dimColor, range: range)
-    }
-
-    return NSAttributedString(attributedString: storage)
-}
-
-// MARK: - Markdown styling extension
-
-extension LinkAwareTextView {
-    func clearPendingWikilinkInsertion() {
-        pendingWikilinkAlias = nil
-        pendingWikilinkSelectionRange = nil
-    }
-
-    func setPlainText(_ plain: String) {
-        guard let storage = textStorage else { return }
-        // Tear down any in-flight/pending AI session BEFORE the storage is
-        // replaced — stale ranges would corrupt the new note or crash on
-        // accept/reject, and the floating bar would linger over new content.
-        teardownAISession()
-        // Stale ranges from a previous file would crash reapplySearchHighlights
-        lastSearchHighlightRanges = []
-        lastSearchFocusIndex = -1
-        // New content invalidates the revealed-block gate so the next caret move
-        // re-evaluates against the new document.
-        lastRevealedBlockRange = nil
-        storage.beginEditing()
-        storage.setAttributedString(NSAttributedString(string: plain))
-        storage.endEditing()
-        applyMarkdownStyling(deferRedraw: !isEditable)
-        if !isEditable {
-            applyPreviewStyling(editingSessionOpen: true)
-        }
-        // Note: hideMarkdownWhileEditing in editable mode is handled in the
-        // Coordinator's styling callback and updateNSView, which have access to appState.
-    }
-
-    /// Called after applyMarkdownStyling() in view/preview mode.
-    /// Hides markdown syntax tokens (delimiters, sigils, fences) by setting
-    /// their font size to near-zero and foreground color to clear, so only the
-    /// styled content is visible.
-    func applyPreviewStyling(document: MarkdownDocument? = nil, refreshPlan: MarkdownEditorRefreshPlan = .fullDocument, editingSessionOpen: Bool = false) {
-        guard let storage = textStorage else { return }
-        let fullRange = NSRange(location: 0, length: storage.length)
-        guard fullRange.length > 0 else { return }
-        let text = storage.string
-        let parsedDocument = document ?? MarkdownDocumentParser().parse(text)
-        let previewSemanticHiding = MarkdownPreviewSemanticHiding.make(from: parsedDocument, isEditable: isEditable)
-        let scopeRange = refreshPlan.affectedRange ?? fullRange
-        let searchRange = (text as NSString).lineRange(for: scopeRange)
-        let fencedCodeBlockRanges = parsedDocument.blocks.compactMap { block -> NSRange? in
-            if case .fencedCodeBlock = block.kind {
-                return block.range
-            }
-            return nil
-        }
-
-        let hiddenAttrs: [NSAttributedString.Key: Any] = [
-            .font: NSFont.systemFont(ofSize: 0.001),
-            .foregroundColor: NSColor.clear,
-        ]
-
-        func hide(_ pattern: String, options: NSRegularExpression.Options = []) {
-            guard let regex = cachedRegex(pattern, options: options) else { return }
-            regex.enumerateMatches(in: text, options: [], range: searchRange) { match, _, _ in
-                guard let range = match?.range else { return }
-                storage.addAttributes(hiddenAttrs, range: range)
-            }
-        }
-
-        func hideGroup(_ pattern: String, group: Int, options: NSRegularExpression.Options = []) {
-            guard let regex = cachedRegex(pattern, options: options) else { return }
-            regex.enumerateMatches(in: text, options: [], range: searchRange) { match, _, _ in
-                guard let match, match.numberOfRanges > group else { return }
-                let r = match.range(at: group)
-                guard r.location != NSNotFound else { return }
-                storage.addAttributes(hiddenAttrs, range: r)
-            }
-        }
-
-        func isInsideFencedCodeBlock(_ range: NSRange) -> Bool {
-            fencedCodeBlockRanges.contains { blockRange in
-                NSIntersectionRange(blockRange, range).length > 0
-            }
-        }
-
-        // applyMarkdownStyling() already ran before this and applied all fonts.
-        // We only need to hide the markdown syntax tokens here.
-        // Do NOT re-apply base fonts — that would undo the heading sizes set by applyMarkdownStyling.
-
-        if !editingSessionOpen {
-            storage.beginEditing()
-        }
-
-        for range in previewSemanticHiding.hiddenRanges where NSIntersectionRange(range, scopeRange).length > 0 {
-            storage.addAttributes(hiddenAttrs, range: range)
-        }
-
-        for block in parsedDocument.blocks {
-            guard case .fencedCodeBlock = block.kind else { continue }
-            guard NSIntersectionRange(block.range, searchRange).length > 0 else { continue }
-
-            let firstLineRange = (text as NSString).lineRange(for: NSRange(location: block.range.location, length: 0))
-            let lastLineLocation = block.range.location + block.range.length - 1
-            let lastLineRange = (text as NSString).lineRange(for: NSRange(location: lastLineLocation, length: 0))
-
-            for lineRange in [firstLineRange, lastLineRange] {
-                let paragraphStyle = (storage.attribute(.paragraphStyle, at: lineRange.location, effectiveRange: nil) as? NSParagraphStyle)?.mutableCopy() as? NSMutableParagraphStyle ?? NSMutableParagraphStyle()
-                paragraphStyle.minimumLineHeight = 0
-                paragraphStyle.maximumLineHeight = 0
-                paragraphStyle.lineSpacing = 0
-                storage.addAttribute(.paragraphStyle, value: paragraphStyle, range: lineRange)
-            }
-        }
-
-        // Bold **text** — hide the ** delimiters
-        hideGroup("(\\*\\*)(.+?)(\\*\\*)", group: 1)
-        hideGroup("(\\*\\*)(.+?)(\\*\\*)", group: 3)
-        // Bold __text__ — hide the __ delimiters
-        hideGroup("(__)(.+?)(__)", group: 1)
-        hideGroup("(__)(.+?)(__)", group: 3)
-
-        // Italic *text* — hide the * delimiters (not **)
-        hideGroup("(?<!\\*)(\\*)(?!\\*)(.+?)(?<!\\*)(\\*)(?!\\*)", group: 1)
-        hideGroup("(?<!\\*)(\\*)(?!\\*)(.+?)(?<!\\*)(\\*)(?!\\*)", group: 3)
-
-        // Inline code `code` — hide the backtick delimiters
-        if let regex = cachedRegex("(`)((?:[^`\\n])+)(`)") {
-            regex.enumerateMatches(in: text, options: [], range: searchRange) { match, _, _ in
-                guard let match, match.numberOfRanges > 3 else { return }
-                let openRange = match.range(at: 1)
-                let closeRange = match.range(at: 3)
-                guard openRange.location != NSNotFound, closeRange.location != NSNotFound else { return }
-                if isInsideFencedCodeBlock(match.range(at: 0)) {
-                    return
-                }
-                storage.addAttributes(hiddenAttrs, range: openRange)
-                storage.addAttributes(hiddenAttrs, range: closeRange)
-            }
-        }
-
-        // Image embeds ![caption](url) — hide ![ and ](url), keep caption visible.
-        // Only hide when caption is non-empty; if [] leave the full markdown visible.
-        hideGroup("(!\\[)([^\\]]+)(\\]\\([^)]+\\))", group: 1)
-        hideGroup("(!\\[)([^\\]]+)(\\]\\([^)]+\\))", group: 3)
-
-        // Dim caption text for image embeds
-        let imageCaptionRegex = cachedRegex("!\\[([^\\]]+)\\]\\([^)]+\\)")
-        imageCaptionRegex?.enumerateMatches(in: text, options: [], range: searchRange) { match, _, _ in
-            guard let match, match.numberOfRanges > 1 else { return }
-            let captionRange = match.range(at: 1)
-            guard captionRange.location != NSNotFound else { return }
-            storage.addAttributes([
-                .foregroundColor: MarkdownTheme.dimColor,
-            ], range: captionRange)
-        }
-
-        storage.endEditing()
-        requestImmediateRedraw(for: scopeRange)
-        lastAppliedEditorDisplayMode = .preview
-        refreshTaskCheckboxButtons()
-
-        // After hiding, reveal the wikilink/image embed the cursor is currently inside.
-        // NOTE: the *block* reveal is intentionally NOT triggered here. applyPreviewStyling
-        // is a pure "hide everything" pass (also used for read-only rendering and initial
-        // load, where the caret sits at 0 inside the first block). Block reveal is a
-        // response to caret movement and is driven from textViewDidChangeSelection instead.
-        if isEditable {
-            revealSemanticInlineMarkdownAtCursor()
-            revealCalloutHeaderAtCursor(document: parsedDocument)
-        }
-    }
-
-    private func revealCalloutHeaderAtCursor(document: MarkdownDocument? = nil) {
-        guard let storage = textStorage else { return }
-        let cursor = selectedRange().location
-        guard cursor != NSNotFound else { return }
-        let parsedDocument = document ?? MarkdownDocumentParser().parse(storage.string)
-        let callouts = parsedDocument.blocks.compactMap { MarkdownCalloutDetector.detect(in: $0, source: parsedDocument.source) }
-        guard let callout = callouts.first(where: { NSLocationInRange(cursor, $0.headerRange) }) else { return }
-
-        // Configured body font, not the fixed 15pt legacy constant (see
-        // revealSemanticInlineMarkdownAtCursor for why).
-        let visibleAttrs: [NSAttributedString.Key: Any] = [
-            .font: settings != nil ? MarkdownTheme.bodyFont(for: settings!) : MarkdownTheme.body,
-            .foregroundColor: MarkdownTheme.dimColor,
-        ]
-        storage.beginEditing()
-        storage.addAttributes(visibleAttrs, range: callout.headerRange)
-        storage.endEditing()
-    }
-
-    func revealSemanticInlineMarkdownAtCursor() {
-        guard let storage = textStorage else { return }
-        let reveal = MarkdownPreviewCursorReveal.make(
-            from: storage.string,
-            cursorLocation: selectedRange().location,
-            isEditable: isEditable
-        )
-        guard !reveal.revealedRanges.isEmpty else { return }
-
-        // Use the configured body font, NOT the fixed 15pt legacy MarkdownTheme.body —
-        // otherwise revealing a token shrinks it whenever the user's editor font ≠ 15.
-        let visibleAttrs: [NSAttributedString.Key: Any] = [
-            .font: settings != nil ? MarkdownTheme.bodyFont(for: settings!) : MarkdownTheme.body,
-            .foregroundColor: MarkdownTheme.dimColor,
-        ]
-
-        storage.beginEditing()
-        for range in reveal.revealedRanges {
-            storage.addAttributes(visibleAttrs, range: range)
-        }
-        storage.endEditing()
-    }
-
-    /// Reveals the raw markdown syntax (dimmed) for the entire parsed block the caret
-    /// is in, so editing always shows the syntax for the block being edited. Re-hiding
-    /// of the block the caret *left* is handled by the next full applyPreviewStyling pass.
-    /// No-ops when the caret stays within the same block as the previous call.
-    func revealCurrentBlockMarkdownAtCursor(document: MarkdownDocument? = nil) {
-        guard isEditable, let storage = textStorage else { return }
-        let cursor = selectedRange().location
-        // The optional `document` lets callers avoid an extra parse when they already
-        // hold one (its `source` is authoritative); otherwise read the live storage.
-        let source = document?.source ?? storage.string
-        let reveal = MarkdownPreviewBlockReveal.make(from: source, cursorLocation: cursor, isEditable: isEditable)
-
-        // Block-change gating: if the caret is still in the same block we revealed last
-        // time, there is nothing new to reveal.
-        if let last = lastRevealedBlockRange, let current = reveal.blockRange, NSEqualRanges(last, current) {
-            return
-        }
-        lastRevealedBlockRange = reveal.blockRange
-
-        guard !reveal.revealedRanges.isEmpty else { return }
-
-        // The hidden delimiters were zeroed to systemFont(0.001); restore a visible
-        // body-sized font and dim color. Body font reads cleanly for every delimiter
-        // kind (**, *, `, [, ]], #, ```); surrounding content keeps its own font from
-        // applyMarkdownStyling.
-        let revealFont = settings != nil ? MarkdownTheme.bodyFont(for: settings!) : MarkdownTheme.body
-
-        storage.beginEditing()
-        for range in reveal.revealedRanges {
-            let safeLoc = max(0, min(range.location, storage.length))
-            let safeLen = min(range.length, storage.length - safeLoc)
-            guard safeLen > 0 else { continue }
-            let safeRange = NSRange(location: safeLoc, length: safeLen)
-            storage.addAttributes([
-                .font: revealFont,
-                .foregroundColor: MarkdownTheme.dimColor,
-            ], range: safeRange)
-        }
-        storage.endEditing()
-        requestImmediateRedraw(for: reveal.blockRange ?? NSRange(location: cursor, length: 0))
-    }
-
-    func applyMarkdownStyling(document: MarkdownDocument? = nil, refreshPlan: MarkdownEditorRefreshPlan = .fullDocument, deferRedraw: Bool = false) {
-        guard let storage = textStorage else { return }
-        let fullRange = NSRange(location: 0, length: storage.length)
-        guard fullRange.length > 0 else {
-            lastAppliedEditorFontSignature = EditorFontSignature(settings: settings)
-            lastAppliedEditorDisplayMode = .markdown
-            clearInlineImagePreviews()
-            clearTaskCheckboxButtons()
-            for key in Array(collapsibleToggleButtons.keys) {
-                collapsibleToggleButtons[key]?.removeFromSuperview()
-            }
-            collapsibleToggleButtons.removeAll()
-            return
-        }
-        let text = storage.string as NSString
-        let parsedDocument = document ?? MarkdownDocumentParser().parse(storage.string)
-        let scopeRange = refreshPlan.affectedRange ?? fullRange
-        let searchRange = text.lineRange(for: scopeRange)
-        lastAppliedEditorDisplayMode = .markdown
-        clearTaskCheckboxButtons()
-        let semanticStyles = MarkdownEditorSemanticStyles.make(from: parsedDocument)
-        let inlineSemanticStyles = MarkdownEditorInlineSemanticStyles.make(from: parsedDocument)
-
-        storage.beginEditing()
-
-        // Use settings-based fonts if available, otherwise fall back to defaults
-        let bodyFont = settings != nil ? MarkdownTheme.bodyFont(for: settings!) : MarkdownTheme.body
-        let monoFont = settings != nil ? MarkdownTheme.monoFont(for: settings!) : MarkdownTheme.mono
-        let h1Font = settings != nil ? MarkdownTheme.h1Font(for: settings!) : MarkdownTheme.h1
-        let h2Font = settings != nil ? MarkdownTheme.h2Font(for: settings!) : MarkdownTheme.h2
-        let h3Font = settings != nil ? MarkdownTheme.h3Font(for: settings!) : MarkdownTheme.h3
-        let h4Font = settings != nil ? MarkdownTheme.h4Font(for: settings!) : MarkdownTheme.h4
-        let boldFont = settings != nil ? MarkdownTheme.boldFont(for: settings!) : NSFont.systemFont(ofSize: 15, weight: .bold)
-        let italicFont = settings != nil ? MarkdownTheme.italicFont(for: settings!) : {
-            let desc = MarkdownTheme.body.fontDescriptor.withSymbolicTraits(.italic)
-            return NSFont(descriptor: desc, size: 15) ?? MarkdownTheme.body
-        }()
-        let lineHeightMultiple = settings != nil ? MarkdownTheme.lineHeightMultiple(for: settings!) : 1.6
-        let baseParagraphStyle = MarkdownTheme.paragraphStyle(font: bodyFont, lineHeightMultiple: lineHeightMultiple)
-
-        storage.setAttributes([
-            .font: bodyFont,
-            .foregroundColor: SynapseTheme.editorForeground,
-            .paragraphStyle: baseParagraphStyle,
-        ], range: scopeRange)
-
-        for heading in semanticStyles.headings {
-            guard NSIntersectionRange(heading.range, scopeRange).length > 0 else { continue }
-            let font: NSFont
-            switch heading.level {
-            case 1: font = h1Font
-            case 2: font = h2Font
-            case 3: font = h3Font
-            default: font = h4Font
-            }
-            let headingParaStyle = MarkdownTheme.paragraphStyle(font: font, lineHeightMultiple: lineHeightMultiple)
-            storage.addAttributes([.font: font, .paragraphStyle: headingParaStyle], range: heading.range)
-            storage.addAttribute(.foregroundColor, value: MarkdownTheme.dimColor, range: heading.markerRange)
-        }
-
-        // Italic first, bold second — bold must win on **word** spans.
-        // The single-star italic regex would otherwise match the inner *word* of **word**
-        // and overwrite the bold font after it was applied.
-        applyRegex("(?<!\\*)\\*(?!\\*)(.+?)(?<!\\*)\\*(?!\\*)", to: text, storage: storage, searchRange: searchRange) { range in
-            storage.addAttribute(.font, value: italicFont, range: range)
-            dimDelimiters(storage: storage, outerRange: range, delimLen: 1)
-        }
-        applyRegex("\\*\\*(.+?)\\*\\*", to: text, storage: storage, searchRange: searchRange) { range in
-            storage.addAttribute(.font, value: boldFont, range: range)
-            dimDelimiters(storage: storage, outerRange: range, delimLen: 2)
-        }
-        // Single-underscore italic before double-underscore bold, same reason as * vs **.
-        // Word-boundary guards prevent matching inside identifiers like snake_case.
-        applyRegex("(?<![\\w_])_(?!_)(.+?)(?<!_)_(?![\\w_])", to: text, storage: storage, searchRange: searchRange) { range in
-            storage.addAttribute(.font, value: italicFont, range: range)
-            dimDelimiters(storage: storage, outerRange: range, delimLen: 1)
-        }
-        applyRegex("__(.+?)__", to: text, storage: storage, searchRange: searchRange) { range in
-            storage.addAttribute(.font, value: boldFont, range: range)
-            dimDelimiters(storage: storage, outerRange: range, delimLen: 2)
-        }
-        // Bold+italic last so it overrides the bold/italic font applied on inner substrings.
-        let boldItalicFont = settings != nil ? MarkdownTheme.boldItalicFont(for: settings!) : {
-            let desc = NSFont.systemFont(ofSize: 15, weight: .bold).fontDescriptor.withSymbolicTraits([.bold, .italic])
-            return NSFont(descriptor: desc, size: 15) ?? NSFont.systemFont(ofSize: 15, weight: .bold)
-        }()
-        applyRegex("\\*\\*\\*(.+?)\\*\\*\\*", to: text, storage: storage, searchRange: searchRange) { range in
-            storage.addAttribute(.font, value: boldItalicFont, range: range)
-            dimDelimiters(storage: storage, outerRange: range, delimLen: 3)
-        }
-        // Strikethrough: ~~text~~ and single ~text~ (with guards so it doesn't hit ~/home or ~~~).
-        applyRegex("~~(.+?)~~", to: text, storage: storage, searchRange: searchRange) { range in
-            storage.addAttribute(.strikethroughStyle, value: NSUnderlineStyle.single.rawValue, range: range)
-            dimDelimiters(storage: storage, outerRange: range, delimLen: 2)
-        }
-        applyRegex("(?<![\\w~])~(?!~)(.+?)(?<!~)~(?![\\w~])", to: text, storage: storage, searchRange: searchRange) { range in
-            storage.addAttribute(.strikethroughStyle, value: NSUnderlineStyle.single.rawValue, range: range)
-            dimDelimiters(storage: storage, outerRange: range, delimLen: 1)
-        }
-        applyRegex("`([^`\\n]+)`", to: text, storage: storage, searchRange: searchRange) { range in
-            storage.addAttributes([.font: monoFont, .backgroundColor: MarkdownTheme.codeBackground], range: range)
-        }
-        let codePad: CGFloat = 10
-        let isDarkMode = NSApp.effectiveAppearance.bestMatch(from: [.darkAqua, .aqua]) == .darkAqua
-        for block in parsedDocument.blocks {
-            guard case let .fencedCodeBlock(_, infoString) = block.kind else { continue }
-            guard NSIntersectionRange(block.range, scopeRange).length > 0 else { continue }
-
-            storage.addAttributes([
-                .font: monoFont,
-                .backgroundColor: MarkdownTheme.codeBackground,
-                .foregroundColor: SynapseTheme.editorForeground,
-                // Marker read by drawBackground(in:) to extend the fill to full width.
-                .codeBlockFullWidthBackground: MarkdownTheme.codeBackground,
-            ], range: block.range)
-
-            if SyntaxHighlighter.isSupportedLanguage(infoString) {
-                SyntaxHighlighter.apply(
-                    to: storage,
-                    codeRange: block.contentRange,
-                    language: infoString,
-                    baseFont: monoFont,
-                    isDarkMode: isDarkMode
-                )
-            }
-
-            // Add bottom padding to the closing fence line so the code block has breathing room
-            // and the copy button has space to sit in.
-            let nsStr = text as NSString
-            let firstLineRange = nsStr.lineRange(for: NSRange(location: block.range.location, length: 0))
-            let firstParaStyle = (storage.attribute(.paragraphStyle, at: firstLineRange.location, effectiveRange: nil) as? NSParagraphStyle)?.mutableCopy() as? NSMutableParagraphStyle ?? NSMutableParagraphStyle()
-            firstParaStyle.paragraphSpacingBefore = 0
-            storage.addAttribute(.paragraphStyle, value: firstParaStyle, range: firstLineRange)
-            // Last line of block → paragraphSpacing (after) and full-width background
-            let lastLineRange = nsStr.lineRange(for: NSRange(location: block.range.location + block.range.length - 1, length: 0))
-            let lastParaStyle = (storage.attribute(.paragraphStyle, at: lastLineRange.location, effectiveRange: nil) as? NSParagraphStyle)?.mutableCopy() as? NSMutableParagraphStyle ?? NSMutableParagraphStyle()
-            lastParaStyle.paragraphSpacing = codePad
-            lastParaStyle.tailIndent = 0
-            lastParaStyle.lineBreakMode = .byWordWrapping
-            storage.addAttribute(.paragraphStyle, value: lastParaStyle, range: lastLineRange)
-        }
-        for block in parsedDocument.blocks {
-            guard case .table = block.kind else { continue }
-            guard NSIntersectionRange(block.range, scopeRange).length > 0 else { continue }
-            storage.addAttribute(.font, value: monoFont, range: block.range)
-        }
-        let calloutRanges = Set(semanticStyles.callouts.map { "\($0.range.location):\($0.range.length)" })
-        for range in semanticStyles.blockquotes {
-            guard !calloutRanges.contains("\(range.location):\(range.length)") else { continue }
-            guard NSIntersectionRange(range, scopeRange).length > 0 else { continue }
-            storage.addAttribute(.foregroundColor, value: MarkdownTheme.dimColor, range: range)
-            // Indent the text so a colored accent bar can live in the gutter without
-            // overlapping the glyphs. drawBackground(in:) paints the bar.
-            let existing = storage.attribute(.paragraphStyle, at: range.location, effectiveRange: nil) as? NSParagraphStyle
-            let paraStyle = (existing?.mutableCopy() as? NSMutableParagraphStyle) ?? NSMutableParagraphStyle()
-            paraStyle.firstLineHeadIndent = 16
-            paraStyle.headIndent = 16
-            storage.addAttribute(.paragraphStyle, value: paraStyle, range: range)
-            storage.addAttribute(.blockquoteLeftBorder, value: MarkdownTheme.linkColor, range: range)
-        }
-        for callout in semanticStyles.callouts {
-            guard NSIntersectionRange(callout.range, scopeRange).length > 0 else { continue }
-            let background = MarkdownTheme.codeBackground.blended(withFraction: 0.2, of: MarkdownTheme.linkColor) ?? MarkdownTheme.codeBackground
-            storage.addAttributes([
-                .backgroundColor: background,
-                .foregroundColor: SynapseTheme.editorForeground,
-            ], range: callout.range)
-            storage.addAttribute(.foregroundColor, value: MarkdownTheme.dimColor, range: callout.markerRange)
-            if let titleRange = callout.titleRange {
-                storage.addAttributes([
-                    .font: boldFont,
-                    .foregroundColor: MarkdownTheme.linkColor,
-                ], range: titleRange)
-            }
-        }
-        if let frontmatter = semanticStyles.frontmatter {
-            if NSIntersectionRange(frontmatter.contentRange, scopeRange).length > 0 {
-                // Use a static/fixed line height for frontmatter that doesn't change with user settings
-                let frontmatterFont = NSFont.systemFont(ofSize: 11)
-                let frontmatterParagraphStyle = MarkdownTheme.paragraphStyle(font: frontmatterFont, lineHeightMultiple: 1.2)
-                storage.addAttributes([
-                    .font: frontmatterFont,
-                    .foregroundColor: SynapseTheme.editorMuted,
-                    .paragraphStyle: frontmatterParagraphStyle,
-                ], range: frontmatter.contentRange)
-            }
-            let openingFence = NSRange(location: frontmatter.range.location, length: min(3, frontmatter.range.length))
-            let closingFence = NSRange(location: frontmatter.range.location + frontmatter.range.length - 3, length: min(3, frontmatter.range.length))
-            if NSIntersectionRange(openingFence, scopeRange).length > 0 {
-                storage.addAttribute(.foregroundColor, value: MarkdownTheme.dimColor, range: openingFence)
-            }
-            if NSIntersectionRange(closingFence, scopeRange).length > 0 {
-                storage.addAttribute(.foregroundColor, value: MarkdownTheme.dimColor, range: closingFence)
-            }
-        }
-        AppState.inlineTagMatches(in: storage.string).forEach { match in
-            guard NSIntersectionRange(match.range, scopeRange).length > 0 else { return }
-            storage.addAttribute(.foregroundColor, value: MarkdownTheme.tagColor, range: match.range)
-            storage.addAttribute(.tagTarget, value: match.normalized, range: match.range)
-        }
-        let noteNames = Set(allFiles.map { $0.deletingPathExtension().lastPathComponent.lowercased() })
-        for entry in inlineSemanticStyles.entries {
-            guard NSIntersectionRange(entry.range, scopeRange).length > 0 || NSIntersectionRange(entry.contentRange, scopeRange).length > 0 else { continue }
-            switch entry.kind {
-            case let .embed(rawTarget):
-                storage.addAttributes([
-                    .foregroundColor: MarkdownTheme.dimColor,
-                    .link: rawTarget,
-                ], range: entry.range)
-            case let .wikiLink(rawTarget, destination, _):
-                let baseName = destination
-                    .components(separatedBy: "#").first?
-                    .trimmingCharacters(in: .whitespaces) ?? destination
-                let resolved = !noteNames.isEmpty && noteNames.contains(baseName.lowercased())
-                storage.addAttributes([
-                    .foregroundColor: resolved ? MarkdownTheme.linkColor : MarkdownTheme.unresolvedLinkColor,
-                    .underlineStyle: NSUnderlineStyle.single.rawValue,
-                    .wikilinkTarget: rawTarget,
-                ], range: entry.range)
-            case let .markdownLink(destination):
-                storage.addAttribute(.foregroundColor, value: MarkdownTheme.dimColor, range: entry.range)
-                storage.addAttributes([
-                    .foregroundColor: MarkdownTheme.linkColor,
-                    .underlineStyle: NSUnderlineStyle.single.rawValue,
-                ], range: entry.contentRange)
-
-                if let url = URL(string: destination), url.scheme != nil {
-                    storage.addAttribute(.link, value: url, range: entry.range)
-                }
-            case .highlight:
-                storage.addAttribute(.backgroundColor, value: NSColor.systemYellow.withAlphaComponent(0.3), range: entry.contentRange)
-            }
-        }
-
-        if let bareURLRegex = LinkAwareTextView.bareURLRegex {
-            bareURLRegex.enumerateMatches(in: storage.string, options: [], range: searchRange) { match, _, _ in
-                guard let match else { return }
-                let range = match.range
-                guard range.location != NSNotFound, range.length > 0 else { return }
-
-                if storage.attribute(.link, at: range.location, effectiveRange: nil) != nil {
-                    return
-                }
-
-                let rawURL = text.substring(with: range)
-                guard let url = URL(string: rawURL) else { return }
-
-                storage.addAttributes([
-                    .foregroundColor: MarkdownTheme.linkColor,
-                    .underlineStyle: NSUnderlineStyle.single.rawValue,
-                    .link: url,
-                ], range: range)
-            }
-        }
-        for range in semanticStyles.thematicBreaks {
-            guard NSIntersectionRange(range, scopeRange).length > 0 else { continue }
-            storage.addAttribute(.foregroundColor, value: MarkdownTheme.dimColor, range: range)
-        }
-
-        // Image embeds are now shown only in sidebar, not inline
-        // Skip adding paragraph spacing for inline image previews
-        /*
-        for match in self.visibleInlineImageMatches() {
-            let paragraphStyle = (storage.attribute(.paragraphStyle, at: match.paragraphRange.location, effectiveRange: nil) as? NSMutableParagraphStyle)
-                ?? NSMutableParagraphStyle()
-            let updatedStyle = paragraphStyle.mutableCopy() as? NSMutableParagraphStyle ?? NSMutableParagraphStyle()
-            updatedStyle.paragraphSpacing = max(updatedStyle.paragraphSpacing, self.inlinePreviewHeight(for: match.source))
-            storage.addAttribute(.paragraphStyle, value: updatedStyle, range: match.paragraphRange)
-            storage.addAttribute(.foregroundColor, value: MarkdownTheme.dimColor, range: match.range)
-        }
-        */
-
-        // Restore Apple Color Emoji on emoji characters after ALL font-setting passes.
-        // Moving this here (rather than immediately after the blanket setAttributes reset)
-        // prevents heading/bold/italic styling passes from overwriting the emoji font,
-        // which was the root cause of the emoji flicker during typing.
-        restoreEmojiFonts(in: storage, range: scopeRange, bodyFont: bodyFont)
-
-        applyCollapsibleStyling(storage: storage)
-        if !deferRedraw {
-            storage.endEditing()
-        }
-        lastAppliedEditorFontSignature = EditorFontSignature(settings: settings)
-        if !deferRedraw {
-            requestImmediateRedraw(for: scopeRange)
-        }
-        reapplySearchHighlights()
-        DispatchQueue.main.async { [weak self] in
-            self?.refreshInlineImagePreviews()
-            self?.refreshCollapsibleToggles()
-            self?.refreshCodeBlockCopyButtons()
-            self?.refreshAISparkle()
-        }
-    }
-
-    // Compiled-once regex cache keyed by "pattern|options.rawValue"
-    private static var regexCache: [String: NSRegularExpression] = [:]
-    private static let bareURLRegex = try? NSRegularExpression(pattern: #"https?://[^"]+?(?=[\s)\]>]|$)"#)
-
-    private func applyRegex(_ pattern: String, to text: NSString, storage _: NSTextStorage, options: NSRegularExpression.Options = [], searchRange: NSRange? = nil, apply: (NSRange) -> Void) {
-        let cacheKey = "\(pattern)|\(options.rawValue)"
-        let regex: NSRegularExpression
-        if let cached = LinkAwareTextView.regexCache[cacheKey] {
-            regex = cached
-        } else if let compiled = try? NSRegularExpression(pattern: pattern, options: options) {
-            LinkAwareTextView.regexCache[cacheKey] = compiled
-            regex = compiled
-        } else {
-            return
-        }
-        let range = searchRange ?? NSRange(location: 0, length: text.length)
-        regex.enumerateMatches(in: text as String, options: [], range: range) { match, _, _ in
-            guard let range = match?.range else { return }
-            apply(range)
-        }
-    }
-
-    /// Re-apply Apple Color Emoji to emoji characters after a blanket font reset.
-    /// `NSTextStorage.setAttributes` replaces the font on every character,
-    /// including emoji — which need the Apple Color Emoji font to render.
-    /// Without this pass emoji momentarily show a fallback glyph (`` ` ``) until
-    /// Core Text resolves the substitution, causing visible flicker.
-    private func restoreEmojiFonts(in storage: NSTextStorage, range: NSRange, bodyFont: NSFont) {
-        let text = storage.string
-        let nsRange = Range(range, in: text)
-        guard let nsRange else { return }
-        let emojiFont = NSFont(name: "Apple Color Emoji", size: bodyFont.pointSize)
-            ?? NSFont.systemFont(ofSize: bodyFont.pointSize)
-
-        // Walk composed character sequences; only touch those containing emoji scalars.
-        var idx = nsRange.lowerBound
-        while idx < nsRange.upperBound {
-            let next = text.index(after: idx)
-            // rangeOfComposedCharacterSequence gives us the full cluster
-            let cluster = text[idx..<next]
-            let isEmoji = cluster.unicodeScalars.contains { scalar in
-                scalar.properties.isEmoji && scalar.value > 0x23F // skip small ASCII-range symbols like #, *, 0-9
-            }
-            if isEmoji {
-                let charRange = NSRange(idx..<next, in: text)
-                storage.addAttribute(.font, value: emojiFont, range: charRange)
-            }
-            idx = next
-        }
-    }
-
-    private func dimDelimiters(storage: NSTextStorage, outerRange: NSRange, delimLen: Int) {
-        guard outerRange.length >= delimLen * 2 else { return }
-        storage.addAttribute(.foregroundColor, value: MarkdownTheme.dimColor, range: NSRange(location: outerRange.location, length: delimLen))
-        storage.addAttribute(.foregroundColor, value: MarkdownTheme.dimColor, range: NSRange(location: outerRange.location + outerRange.length - delimLen, length: delimLen))
-    }
-
-    private func requestImmediateRedraw(for range: NSRange) {
-        guard range.length > 0 else { return }
-        if let layoutManager, let textContainer {
-            layoutManager.invalidateDisplay(forCharacterRange: range)
-            layoutManager.ensureLayout(for: textContainer)
-            var redrawRect = layoutManager.boundingRect(forGlyphRange: layoutManager.glyphRange(forCharacterRange: range, actualCharacterRange: nil), in: textContainer)
-            redrawRect.origin.x += textContainerOrigin.x
-            redrawRect.origin.y += textContainerOrigin.y
-            if !redrawRect.isEmpty {
-                setNeedsDisplay(redrawRect.insetBy(dx: -24, dy: -24))
-            }
-        }
-        needsDisplay = true
-        if range.length == (textStorage?.length ?? 0) {
-            setNeedsDisplay(bounds)
-        }
-    }
-
-    fileprivate func shouldSkipIncrementalMarkdownRestyle(
-        document: MarkdownDocument,
-        refreshPlan: MarkdownEditorRefreshPlan,
-        editedRange: NSRange
-    ) -> Bool {
-        guard case let .blockRange(blockRange) = refreshPlan.kind else { return false }
-        guard let block = document.blocks.first(where: { NSEqualRanges($0.range, blockRange) }) else { return false }
-        guard case .paragraph = block.kind, block.inlineTokens.isEmpty else { return false }
-
-        let nsText = string as NSString
-        guard nsText.length > 0 else { return false }
-
-        let probeLocation = min(max(0, editedRange.location), max(0, nsText.length - 1))
-        let probeLength = min(max(1, editedRange.length), nsText.length - probeLocation)
-        let probeRange = NSRange(location: probeLocation, length: probeLength)
-        let probeText = nsText.substring(with: probeRange)
-
-        return !containsMarkdownTrigger(in: probeText)
-    }
-
-    private func containsMarkdownTrigger(in text: String) -> Bool {
-        let triggerCharacters = CharacterSet(charactersIn: "*_`[]!~#>|-:/")
-        return text.rangeOfCharacter(from: triggerCharacters) != nil
-    }
-
-    // MARK: - Collapsible section toggle buttons
-
-    /// Applies collapsed-content hiding to the text storage and positions toggle arrow buttons.
-    /// Must be called from within or after `applyMarkdownStyling` once layout is ready.
-    func applyCollapsibleStyling(storage: NSTextStorage) {
-        guard storage.length > 0 else { return }
-
-        let text = storage.string
-        let sections = collapsibleParser.parse(text)
-        let fileURL = currentFileURL ?? AppConstants.unsavedFileURL
-
-        // When the file has no session state yet, auto-initialise each section:
-        // collapse it if it has >= 10 lines, expand it otherwise.
-        if !collapsibleStateManager.hasSessionState(for: fileURL) {
-            for section in sections {
-                guard section.contentRange.length > 0 else { continue }
-                let shouldCollapse = section.contentLineCount(in: text) >= 10
-                collapsibleStateManager.setCollapsed(shouldCollapse,
-                                                     for: section.getIdentifier(),
-                                                     in: fileURL)
-            }
-        }
-
-        for section in sections {
-            let sectionId = section.getIdentifier()
-            let isCollapsed = collapsibleStateManager.isCollapsed(sectionId, in: fileURL)
-
-            guard section.contentRange.length > 0 else { continue }
-            let contentRange = section.contentRange
-
-            // Safety: clamp to storage length
-            let safeLocation = min(contentRange.location, storage.length)
-            let safeLength = min(contentRange.length, storage.length - safeLocation)
-            guard safeLength > 0 else { continue }
-            let safeRange = NSRange(location: safeLocation, length: safeLength)
-
-            if isCollapsed {
-                // Hide content: make it invisible and zero-height
-                let hiddenStyle = NSMutableParagraphStyle()
-                hiddenStyle.maximumLineHeight = 0.001
-                hiddenStyle.minimumLineHeight = 0.001
-                storage.addAttributes([
-                    .foregroundColor: NSColor.clear,
-                    .font: NSFont.systemFont(ofSize: 0.001),
-                    .paragraphStyle: hiddenStyle,
-                ], range: safeRange)
-            }
-        }
-    }
-
-    /// Positions (or creates) a small arrow toggle button in the left margin of each
-    /// collapsible section header line, and removes buttons for sections that no longer exist.
-    func refreshCollapsibleToggles() {
-        guard let layoutManager, let textContainer else { return }
-        layoutManager.ensureLayout(for: textContainer)
-
-        let text = string
-        let sections = collapsibleParser.parse(text)
-        let fileURL = currentFileURL ?? AppConstants.unsavedFileURL
-
-        let activeKeys = Set(sections.map { $0.getIdentifier() })
-
-        // Remove stale buttons
-        for key in Array(collapsibleToggleButtons.keys) where !activeKeys.contains(key) {
-            collapsibleToggleButtons[key]?.removeFromSuperview()
-            collapsibleToggleButtons.removeValue(forKey: key)
-        }
-
-        for section in sections {
-            guard section.contentRange.length > 0 else {
-                // No indented content — remove button if present
-                let key = section.getIdentifier()
-                collapsibleToggleButtons[key]?.removeFromSuperview()
-                collapsibleToggleButtons.removeValue(forKey: key)
-                continue
-            }
-
-            let sectionId = section.getIdentifier()
-            let isCollapsed = collapsibleStateManager.isCollapsed(sectionId, in: fileURL)
-
-            // Anchor the disclosure control to the list marker itself so it aligns
-            // with the first visible line rather than the broader header range.
-            let markerRange = NSRange(location: section.headerRange.location, length: 1)
-            let markerGlyphRange = layoutManager.glyphRange(forCharacterRange: markerRange, actualCharacterRange: nil)
-            var markerRect = layoutManager.boundingRect(forGlyphRange: markerGlyphRange, in: textContainer)
-            markerRect.origin.x += textContainerOrigin.x
-            markerRect.origin.y += textContainerOrigin.y
-
-            let buttonSize: CGFloat = 28
-            let buttonFrame = collapsibleToggleFrame(
-                forMarkerRect: markerRect,
-                textContainerOrigin: textContainerOrigin,
-                buttonSize: buttonSize
-            )
-
-            let button: CollapsibleToggleButton
-            if let existing = collapsibleToggleButtons[sectionId] {
-                button = existing
-            } else {
-                button = CollapsibleToggleButton(frame: buttonFrame)
-                addSubview(button)
-                collapsibleToggleButtons[sectionId] = button
-            }
-
-            button.isCollapsed = isCollapsed
-            button.frame = buttonFrame
-            button.toolTip = isCollapsed ? "Expand section" : "Collapse section"
-
-            // Use target/action — capture the identifier by value
-            let capturedId = sectionId
-            button.target = self
-            button.action = #selector(collapsibleToggleTapped(_:))
-            button.identifier = NSUserInterfaceItemIdentifier(capturedId)
-        }
-    }
-
-    // MARK: - Inline AI editing
-
-    /// Positions a single reused ✨ button just past the end of the caret's line content
-    /// (or past the selection when text is selected). Anchors to the *used* width of the
-    /// caret's line fragment, so on an empty line it sits next to the caret rather than
-    /// at the far right of the text container. Cheap: one layout lookup, no parsing.
-    func refreshAISparkle() {
-        guard let layoutManager, let textContainer else { return }
-        // Respect the user's show/hide preference (default on).
-        guard settings?.showAISparkle ?? true else {
-            aiSparkleButton?.isHidden = true
-            return
-        }
-        let sel = selectedRange()
-        let ns = string as NSString
-
-        // The character index whose line we anchor to: selection end, or the caret.
-        let anchorIndex = max(0, min(sel.length > 0 ? sel.location + sel.length : sel.location, ns.length))
-
-        let fallbackLineHeight = layoutManager.defaultLineHeight(for: font ?? NSFont.systemFont(ofSize: NSFont.systemFontSize))
-        var lineRect: NSRect
-        if sel.length > 0 {
-            // Non-empty selection: anchor just past the trailing edge of its glyphs.
-            let selGlyphs = layoutManager.glyphRange(forCharacterRange: sel, actualCharacterRange: nil)
-            lineRect = layoutManager.boundingRect(forGlyphRange: selGlyphs, in: textContainer)
-        } else if anchorIndex == ns.length
-                    && (ns.length == 0 || ns.substring(with: NSRange(location: ns.length - 1, length: 1)).rangeOfCharacter(from: .newlines) != nil) {
-            // Caret on the final empty line (empty doc, or after a trailing newline). The
-            // layout manager tracks this as the "extra line fragment". Its used rect is
-            // the caret position on that empty line.
-            let extra = layoutManager.extraLineFragmentUsedRect
-            if extra.height > 0 {
-                lineRect = extra
-            } else {
-                lineRect = NSRect(x: 0, y: 0, width: 0, height: fallbackLineHeight)
-            }
-        } else {
-            // Caret on a non-trailing (possibly empty) line: use that line fragment's USED
-            // rect, whose width reflects the actual typeset content (≈ the caret x on an
-            // empty line), not the full container width — which is what made the ✨ fly
-            // off to the right.
-            let glyphIndex = min(layoutManager.glyphIndexForCharacter(at: anchorIndex), max(0, layoutManager.numberOfGlyphs - 1))
-            lineRect = layoutManager.lineFragmentUsedRect(forGlyphAt: glyphIndex, effectiveRange: nil)
-        }
-
-        var rect = lineRect
-        rect.origin.x += textContainerOrigin.x
-        rect.origin.y += textContainerOrigin.y
-
-        let size: CGFloat = 18
-        let frame = NSRect(x: rect.maxX + 6, y: rect.minY + (rect.height - size) / 2, width: size, height: size)
-
-        let button: AISparkleButton
-        if let existing = aiSparkleButton {
-            button = existing
-        } else {
-            button = AISparkleButton(frame: frame)
-            button.target = self
-            button.action = #selector(aiSparkleTapped)
-            addSubview(button)
-            aiSparkleButton = button
-        }
-        // This runs on every caret move; avoid needless invalidation when nothing moved.
-        if button.frame != frame { button.frame = frame }
-        button.isHidden = (aiBarHostingView != nil)   // hide while the bar is open
-    }
-
-    @objc private func aiSparkleTapped() {
-        let sel = selectedRange()
-        let mode: InlineAIBarMode = sel.length > 0 ? .rewrite : .generate
-        presentAIBar(mode: mode, at: sel)
-    }
-
-    private func presentAIBar(mode: InlineAIBarMode, at sel: NSRange) {
-        dismissAIBar()
-        aiBarOriginalSelection = sel
-        aiBarUserMoved = false
-        aiBarDragStartOrigin = nil
-
-        let defaultModel = AIModel(apiID: settings?.aiDefaultModel ?? AIModel.default.apiID)
-        let model = InlineAIBarModel(mode: mode, model: defaultModel)
-        model.allFiles = aiAppState?.allFiles ?? []
-        model.allFolders = aiAppState?.allFolders() ?? []
-
-        model.onSubmit = { [weak self] prompt, chosen in
-            self?.startAIStream(prompt: prompt, model: chosen, mode: mode, selection: sel)
-        }
-        model.onRetry = { [weak self] prompt, chosen in
-            // Discard the previous output so the re-run replaces it instead of
-            // appending, then stream fresh from the original anchor/selection.
-            self?.inlineAIController.discardOutput()
-            self?.clearAIDiffColors()
-            self?.aiBarModel?.awaitingAcceptReject = false
-            self?.startAIStream(prompt: prompt, model: chosen, mode: mode, selection: sel)
-        }
-        model.onStop   = { [weak self] in self?.stopAIStream() }
-        model.onAccept = { [weak self] in self?.acceptAI() }
-        model.onReject = { [weak self] in self?.rejectAI() }
-        model.onCancel = { [weak self] in self?.dismissAIBar() }
-        model.onDrag = { [weak self] translation in self?.dragAIBar(by: translation) }
-        model.onDragEnded = { [weak self] in self?.aiBarDragStartOrigin = nil }
-        model.onContentSizeMayHaveChanged = { [weak self] in self?.resizeAIBarToFit() }
-        aiBarModel = model
-
-        let host = NSHostingView(rootView: InlineAIBarView(model: model))
-        host.frame = aiBarFrame(below: sel)
-        addSubview(host)
-        aiBarHostingView = host
-        refreshAISparkle()
-    }
-
-    /// Frame for the AI bar. Anchored just below the bottom of the affected region so it
-    /// never overlaps the streamed text/diff (the region is the end of the streamed
-    /// `newRange`/`originalRange` once streaming starts, else the selection/cursor). If
-    /// placing it below would push it past the bottom of the visible viewport (a long
-    /// diff), it is placed ABOVE the top of the affected region instead, so it stays on
-    /// screen and still clears the diff.
-    private func aiBarFrame(below sel: NSRange, size: NSSize? = nil) -> NSRect {
-        guard let layoutManager, let textContainer else { return .zero }
-        let ns = string as NSString
-        let barSize = size ?? aiBarFittedSize()
-        let width = barSize.width
-        let barHeight = barSize.height
-
-        func yOffset(forCharacterIndex index: Int) -> (top: CGFloat, bottom: CGFloat) {
-            let safe = max(0, min(index, ns.length))
-            let gr = layoutManager.glyphRange(forCharacterRange: NSRange(location: safe, length: 0), actualCharacterRange: nil)
-            var r = layoutManager.boundingRect(forGlyphRange: gr, in: textContainer)
-            r.origin.y += textContainerOrigin.y
-            return (r.minY, r.maxY)
-        }
-
-        // Bottom of the affected region (prefer streamed text) and top of it (for the
-        // above-placement fallback).
-        var bottomAnchor = sel.length > 0 ? sel.location + sel.length : sel.location
-        if let nr = inlineAIController.newRange { bottomAnchor = max(bottomAnchor, NSMaxRange(nr)) }
-        if let orig = inlineAIController.originalRange { bottomAnchor = max(bottomAnchor, NSMaxRange(orig)) }
-        let topAnchor = min(sel.location, inlineAIController.originalRange?.location ?? sel.location)
-
-        let belowY = yOffset(forCharacterIndex: bottomAnchor).bottom + 6
-        let visible = enclosingScrollView?.documentVisibleRect ?? visibleRect
-
-        // If the bar placed below would run past the visible area, place it above the region.
-        if belowY + barHeight > visible.maxY {
-            let aboveY = yOffset(forCharacterIndex: topAnchor).top - barHeight - 6
-            let clampedY = max(visible.minY + 6, aboveY)
-            return NSRect(x: 12, y: clampedY, width: width, height: barHeight)
-        }
-        return NSRect(x: 12, y: belowY, width: width, height: barHeight)
-    }
-
-    /// The bar's content-fitted size (drag handle + growing prompt + suggestion list),
-    /// clamped to the editor width and a sane height range. fittingSize needs the
-    /// target width set on the host first.
-    private func aiBarFittedSize() -> NSSize {
-        let width = min(bounds.width - 24, 520)
-        var height: CGFloat = 80
-        if let host = aiBarHostingView {
-            host.frame.size.width = width
-            let fitting = host.fittingSize.height
-            if fitting > 0 { height = max(60, min(fitting, 360)) }
-        }
-        return NSSize(width: width, height: height)
-    }
-
-    /// Re-anchors the bar below the current affected region (called as text streams in
-    /// and when streaming finishes) so it tracks the growing diff instead of covering it.
-    /// No-op once the user has dragged the bar to a manual position.
-    private func repositionAIBar() {
-        guard let host = aiBarHostingView, !aiBarUserMoved else { return }
-        // Streaming doesn't change the bar's content, so reuse its current size —
-        // avoids a full SwiftUI fitting pass per streamed delta.
-        host.frame = aiBarFrame(below: aiBarOriginalSelection, size: host.frame.size)
-    }
-
-    /// Resizes the bar to fit its content (prompt growth, suggestion list). Preserves the
-    /// user-dragged origin if they moved it; otherwise re-anchors below the affected region.
-    private func resizeAIBarToFit() {
-        guard let host = aiBarHostingView else { return }
-        if aiBarUserMoved {
-            host.frame.size = aiBarFittedSize()
-        } else {
-            host.frame = aiBarFrame(below: aiBarOriginalSelection)
-        }
-    }
-
-    /// Moves the bar by a drag-handle translation. The bar is a subview of the text view,
-    /// which is a flipped NSView (y grows downward) — same direction as SwiftUI's global
-    /// translation — so the y delta is ADDED, not negated. The translation is cumulative
-    /// from drag start, so we offset the origin captured when the drag began.
-    private func dragAIBar(by translation: CGSize) {
-        guard let host = aiBarHostingView else { return }
-        aiBarUserMoved = true
-        let start = aiBarDragStartOrigin ?? host.frame.origin
-        if aiBarDragStartOrigin == nil { aiBarDragStartOrigin = start }
-        let newOrigin = NSPoint(x: start.x + translation.width,
-                                y: start.y + translation.height)
-        // Keep the bar within the visible area.
-        let visible = enclosingScrollView?.documentVisibleRect ?? visibleRect
-        let clampedX = min(max(newOrigin.x, visible.minX + 4), visible.maxX - host.frame.width - 4)
-        let clampedY = min(max(newOrigin.y, visible.minY + 4), visible.maxY - host.frame.height - 4)
-        host.frame.origin = NSPoint(x: clampedX, y: clampedY)
-    }
-
-    /// Shared teardown core: cancels any in-flight stream, closes the operation's
-    /// single undo group, and removes the bar.
-    private func cancelAIStreamAndRemoveBar() {
-        aiStreamTask?.cancel(); aiStreamTask = nil
-        endAIUndoGroup()
-        aiBarHostingView?.removeFromSuperview(); aiBarHostingView = nil
-        aiBarModel = nil
-    }
-
-    private func dismissAIBar() {
-        cancelAIStreamAndRemoveBar()
-        refreshAISparkle()
-    }
-
-    /// Tears down any in-flight or pending inline-AI session. Called when the
-    /// document is swapped (note/tab switch) so stale ranges can't corrupt the
-    /// new note or crash on accept/reject. Does NOT touch storage: the
-    /// about-to-run setPlainText replaces the whole attributed string, so old
-    /// diff colors vanish with it.
-    func teardownAISession() {
-        guard aiBarHostingView != nil || inlineAIController.mode != .idle else { return }
-        cancelAIStreamAndRemoveBar()
-        inlineAIController.resetWithoutMutating()
-    }
-
-    /// Re-applies transient AI diff colors after a styling pass, if a session is active.
-    /// The normal markdown restyle blanket-sets foreground colors, wiping the diff
-    /// colors; this restores them so they don't flicker mid-stream.
-    func reapplyAIDiffColorsIfActive() {
-        guard inlineAIController.mode != .idle else { return }
-        applyAIDiffColors()
-    }
-
-    /// Applies an AI text mutation through the standard NSTextView edit path so it
-    /// registers with the undo manager (Cmd-Z reverts AI insertions/rewrites). Bounds-safe.
-    /// All edits in one AI operation are coalesced into a single undo group (see
-    /// `beginAIUndoGroup`/`endAIUndoGroup`) so one Cmd-Z reverts the whole thing.
-    private func performAIEdit(_ range: NSRange, _ replacement: String) {
-        guard let storage = textStorage else { return }
-        guard range.location >= 0, NSMaxRange(range) <= storage.length else { return }
-        if shouldChangeText(in: range, replacementString: replacement) {
-            replaceCharacters(in: range, with: replacement)
-            didChangeText()
-        }
-    }
-
-    /// Opens an undo group so every streamed delta + the accept/reject deletion collapse
-    /// into a single Cmd-Z. Also disables NSTextView's automatic per-keystroke coalescing
-    /// boundary so the deltas don't split into separate undo steps.
-    private func beginAIUndoGroup() {
-        guard !aiUndoGroupOpen else { return }
-        breakUndoCoalescing()
-        undoManager?.beginUndoGrouping()
-        aiUndoGroupOpen = true
-    }
-
-    /// Closes the AI undo group opened by `beginAIUndoGroup` (idempotent).
-    private func endAIUndoGroup() {
-        guard aiUndoGroupOpen else { return }
-        undoManager?.endUndoGrouping()
-        breakUndoCoalescing()
-        aiUndoGroupOpen = false
-    }
-
-    private func startAIStream(prompt: String, model: AIModel, mode: InlineAIBarMode, selection sel: NSRange) {
-        guard let storage = textStorage else { return }
-        guard let key = KeychainStore().get(), !key.isEmpty else {
-            aiBarModel?.errorMessage = "Add your Anthropic API key in Settings →"
-            return
-        }
-
-        // Reuse the vault lists captured when the bar was presented; allFolders()
-        // walks every file's ancestor chain, so don't recompute it per submit.
-        let resolver = AIContextResolver(
-            allFiles: aiBarModel?.allFiles ?? [],
-            allFolders: aiBarModel?.allFolders ?? [],
-            readContents: { try? String(contentsOf: $0, encoding: .utf8) })
-        let resolved = resolver.resolve(prompt: prompt)
-
-        if mode == .generate {
-            inlineAIController.beginGenerate(in: storage, at: sel.location)
-        } else {
-            inlineAIController.beginRewrite(in: storage, selection: sel)
-        }
-        // Route the controller's text mutations through the undo-registering path so the
-        // whole AI edit is undoable with Cmd-Z (one logical change, not silent storage edits).
-        inlineAIController.performEdit = { [weak self] range, replacement in
-            self?.performAIEdit(range, replacement)
-        }
-        // Group every edit in this operation (all deltas + the accept/reject deletion)
-        // into a single undo step. Idempotent, so Retry re-entry keeps the same group.
-        beginAIUndoGroup()
-
-        let selectionText = mode == .rewrite ? (string as NSString).substring(with: sel) : nil
-        let body = AIRequestBuilder.build(
-            mode: mode,
-            prompt: prompt, noteText: string,
-            selection: selectionText, context: resolved.blocks, model: model)
-
-        aiBarModel?.isStreaming = true
-        if resolved.truncated {
-            aiBarModel?.errorMessage = "Context truncated to fit."
-        } else if !resolved.missing.isEmpty {
-            aiBarModel?.errorMessage = "\(resolved.missing.count) reference(s) not found."
-        } else {
-            aiBarModel?.errorMessage = nil
-        }
-
-        let client = AnthropicClient(apiKey: key)
-        aiStreamTask = Task { [weak self] in
-            do {
-                for try await delta in client.stream(body: body) {
-                    await MainActor.run {
-                        // appendDelta routes through performAIEdit, which calls didChangeText().
-                        self?.inlineAIController.appendDelta(delta)
-                        self?.colorAIDelta(appendedLength: (delta as NSString).length)
-                        self?.repositionAIBar()
-                    }
-                }
-                await MainActor.run { self?.finishAIStream(mode: mode) }
-            } catch {
-                await MainActor.run { self?.handleAIError(error) }
-            }
-        }
-    }
-
-    private func stopAIStream() {
-        aiStreamTask?.cancel(); aiStreamTask = nil
-        // finishAIStream owns the per-mode end-of-session rules (generate: reset +
-        // dismiss; rewrite: await accept/reject).
-        finishAIStream(mode: aiBarModel?.mode ?? .generate)
-    }
-
-    private func finishAIStream(mode: InlineAIBarMode) {
-        aiBarModel?.isStreaming = false
-        if mode == .rewrite {
-            aiBarModel?.awaitingAcceptReject = true
-        } else {
-            inlineAIController.cancel()
-            dismissAIBar()
-        }
-        applyAIDiffColors()
-        repositionAIBar()
-    }
-
-    private func handleAIError(_ error: Error) {
-        aiBarModel?.isStreaming = false
-        if let e = error as? AnthropicClient.ClientError {
-            switch e {
-            case .invalidKey: aiBarModel?.errorMessage = "Invalid API key — check Settings."
-            case .server(let s): aiBarModel?.errorMessage = "Server error (\(s)). Try again."
-            case .badResponse: aiBarModel?.errorMessage = "Unexpected response. Try again."
-            }
-        } else {
-            aiBarModel?.errorMessage = "Network error. Try again."
-        }
-        if aiBarModel?.mode == .generate {
-            inlineAIController.cancel()   // generate: reset to idle so a retry starts clean
-        }
-        if aiBarModel?.mode == .rewrite { aiBarModel?.awaitingAcceptReject = true }
-    }
-
-    /// Shared accept/reject epilogue: resolve the diff via the controller, restore
-    /// normal styling, sync the final text to the binding, and close the bar.
-    private func resolveAIRewrite(_ resolve: () -> Void) {
-        resolve()
-        clearAIDiffColors()
-        didChangeText()
-        dismissAIBar()
-    }
-
-    private func acceptAI() { resolveAIRewrite(inlineAIController.accept) }
-
-    private func rejectAI() { resolveAIRewrite(inlineAIController.reject) }
-
-    /// Colors only the newly appended streamed delta (green) — O(delta) per chunk
-    /// instead of re-coloring the whole accumulated diff (O(total), quadratic over a
-    /// stream). The first delta falls back to the full pass so the original range
-    /// gets its strikethrough/red at the same moment it always has; later wipes by
-    /// styling passes are restored by `reapplyAIDiffColorsIfActive`.
-    private func colorAIDelta(appendedLength: Int) {
-        guard let storage = textStorage,
-              let nr = inlineAIController.newRange, appendedLength > 0 else { return }
-        guard nr.length > appendedLength else {
-            applyAIDiffColors()
-            return
-        }
-        let sub = NSRange(location: NSMaxRange(nr) - appendedLength, length: appendedLength)
-        guard sub.location >= 0, NSMaxRange(sub) <= storage.length else { return }
-        storage.addAttribute(.foregroundColor, value: NSColor.systemGreen, range: sub)
-    }
-
-    private func applyAIDiffColors() {
-        guard let storage = textStorage else { return }
-        if let orig = inlineAIController.originalRange, orig.length > 0,
-           NSMaxRange(orig) <= storage.length {
-            storage.addAttribute(.strikethroughStyle, value: NSUnderlineStyle.single.rawValue, range: orig)
-            storage.addAttribute(.foregroundColor, value: NSColor.systemRed, range: orig)
-        }
-        if let nr = inlineAIController.newRange, nr.length > 0,
-           NSMaxRange(nr) <= storage.length {
-            storage.addAttribute(.foregroundColor, value: NSColor.systemGreen, range: nr)
-        }
-    }
-
-    private func clearAIDiffColors() {
-        guard let storage = textStorage else { return }
-        let full = NSRange(location: 0, length: storage.length)
-        storage.removeAttribute(.strikethroughStyle, range: full)
-        refreshEditorForCurrentDisplayMode(self)
-    }
-
-    @objc private func collapsibleToggleTapped(_ sender: NSControl) {
-        let sectionId = sender.identifier?.rawValue ?? ""
-        guard !sectionId.isEmpty else { return }
-        let fileURL = currentFileURL ?? AppConstants.unsavedFileURL
-        let current = collapsibleStateManager.isCollapsed(sectionId, in: fileURL)
-        collapsibleStateManager.setCollapsed(!current, for: sectionId, in: fileURL)
-        refreshEditorForCurrentDisplayMode(self)
-    }
-
-    private func clearInlineImagePreviews() {
-        for key in Array(inlineImageViews.keys) {
-            inlineImageViews[key]?.removeFromSuperview()
-            inlineImageViews.removeValue(forKey: key)
-        }
-
-        for key in Array(inlineVideoViews.keys) {
-            inlineVideoViews[key]?.removeFromSuperview()
-            inlineVideoViews.removeValue(forKey: key)
-        }
-
-        clearCodeBlockCopyButtons()
-    }
-
-}
-
-#if DEBUG
-private func debugLog(_ msg: String) {
-    let line = "[Synapse] \(msg)\n"
-    if let data = line.data(using: .utf8) {
-        if FileManager.default.fileExists(atPath: "/tmp/Synapse_debug.log") {
-            if let fh = FileHandle(forWritingAtPath: "/tmp/Synapse_debug.log") {
-                fh.seekToEndOfFile()
-                fh.write(data)
-                fh.closeFile()
-            }
-        } else {
-            FileManager.default.createFile(atPath: "/tmp/Synapse_debug.log", contents: data)
-        }
-    }
-
-}
-#else
-@inline(__always) private func debugLog(_ msg: String) {}
-#endif
-
-// MARK: - LinkAwareTextView
-
-class LinkAwareTextView: NSTextView {
-    enum EditorDisplayMode {
-        case markdown
-        case preview
-    }
-
-    var allFiles: [URL] = []
-    var onOpenFile: ((URL, Bool) -> Void)?
-    var onOpenTag: ((String, Bool) -> Void)?  // (tag, openInNewTab)
-    var onActivatePane: (() -> Void)?
-    var onCreateNote: ((String, URL?) -> Void)?  // name, preferred directory
-    var onOpenExternalURL: ((URL) -> Void)?  // External URL opening (defaults to NSWorkspace)
-    var onSelectEmbed: ((String) -> Void)?  // embed ID when clicking on markdown
-    var currentFileURL: URL?
-    var onMatchCountUpdate: ((Int) -> Void)?
-    /// Only the editor participating in global commands (current focused note) should react to
-    /// find/replace notifications that mutate text. Mirrors `onMatchCountUpdate` gating.
-    var participatesInGlobalSearch: Bool = false
-    var onWikiLinkRequest: (() -> Void)?   // Called when [[ is typed
-    var onWikiLinkComplete: ((URL) -> Void)?  // Called when a file is selected for wiki link
-    var onWikiLinkDismiss: (() -> Void)?   // Called when the picker is dismissed via ESC
-    var slashCommandNowProvider: () -> Date = Date.init
-    var slashCommandTimeZone: TimeZone = .current
-    /// Called when CMD-K fires but the editor has no selection, so the normal command palette should open.
-    var onCommandPaletteFallback: (() -> Void)?
-    
-    // Settings manager for font configuration
-    var settings: SettingsManager?
-    fileprivate var lastAppliedEditorFontSignature: EditorFontSignature? = nil
-
-    private var completionPopover: NSPopover?
-    private var completionVC: CompletionViewController?
-    fileprivate var linkTypingRange: NSRange?
-    /// Set when the user ESCs the wiki-link picker. Suppresses reopening the picker
-    /// until the cursor leaves the current [[ token (which calls dismissCompletion).
-    fileprivate var wikilinkPickerSuppressed = false
-    /// Selected text captured before the wikilink palette opens; used to produce [[name|alias]].
-    fileprivate var pendingWikilinkAlias: String? = nil
-    /// Original selection captured before the wikilink palette steals focus.
-    fileprivate var pendingWikilinkSelectionRange: NSRange? = nil
-    var lastAppliedEditorDisplayMode: EditorDisplayMode? = nil
-    private var eventMonitor: Any?
-    private var inlineImageViews: [String: NSImageView] = [:]
-    private var inlineVideoViews: [String: YouTubePreviewView] = [:]
-    private var isPrettifyingTable = false
-
-    // MARK: - Collapsible sections
-    private let collapsibleParser = CollapsibleSectionParser()
-    private let collapsibleStateManager = CollapsibleStateManager()
-    /// Toggle buttons keyed by section identifier ("headerOffset-headerLength")
-    private var collapsibleToggleButtons: [String: CollapsibleToggleButton] = [:]
-
-    // MARK: - Inline AI editing
-    let inlineAIController = InlineAIController()
-    /// True while a rewrite diff is on screen awaiting accept/reject — the buffer
-    /// holds both original and new text, which must not be synced/saved as-is.
-    var hasPendingAIDiff: Bool { inlineAIController.mode == .rewrite }
-    private var aiSparkleButton: AISparkleButton?
-    private var aiBarHostingView: NSHostingView<InlineAIBarView>?
-    private var aiBarModel: InlineAIBarModel?
-    private var aiStreamTask: Task<Void, Never>?
-    /// The selection/cursor captured when the bar opened — used to re-anchor the bar
-    /// below the affected region as text streams in.
-    private var aiBarOriginalSelection: NSRange = NSRange(location: 0, length: 0)
-    /// The bar's origin captured at the start of a drag (nil when not dragging).
-    private var aiBarDragStartOrigin: NSPoint?
-    /// Once the user drags the bar, stop auto-repositioning it below the streamed text.
-    private var aiBarUserMoved = false
-    /// True while an AI undo group is open (so the whole operation undoes as one step).
-    private var aiUndoGroupOpen = false
-    /// Injected at setup; source of vault files for @-context.
-    weak var aiAppState: AppState?
-
-    // MARK: - Embedded Notes (for side panel)
-    private static let embedRegex = try? NSRegularExpression(pattern: #"!\[\[([^\]]+)\]\]"#)
-
-    private static let inlineImageRegex = try? NSRegularExpression(pattern: #"!\[([^\]]*)\]\((.+?)\)"#, options: [])
-
-    /// Extends code-block background fills to the full container width.
-    /// NSAttributedString's .backgroundColor only covers the glyph bounds for that run.
-    /// For the closing fence line the background stops at the last visible glyph,
-    /// leaving a gap on the right. We intercept drawBackground and repaint any run
-    /// that carries the custom `.codeBlockFullWidthBackground` marker attribute as a
-    /// full-width band.
-    override func drawBackground(in rect: NSRect) {
-        super.drawBackground(in: rect)
-
-        guard let storage = textStorage,
-              let layoutManager = layoutManager,
-              let textContainer = textContainer else { return }
-
-        let containerWidth = textContainer.containerSize.width
-        let insetX = textContainerOrigin.x
-        let insetY = textContainerOrigin.y
-
-        var charIndex = 0
-        let length = storage.length
-        while charIndex < length {
-            var effectiveRange = NSRange(location: NSNotFound, length: 0)
-            guard let color = storage.attribute(.codeBlockFullWidthBackground, at: charIndex, effectiveRange: &effectiveRange) as? NSColor,
-                  effectiveRange.location != NSNotFound else {
-                charIndex = effectiveRange.location != NSNotFound ? effectiveRange.location + effectiveRange.length : charIndex + 1
-                continue
-            }
-
-            let glyphRange = layoutManager.glyphRange(forCharacterRange: effectiveRange, actualCharacterRange: nil)
-            var lineStart = glyphRange.location
-            let glyphEnd = glyphRange.location + glyphRange.length
-
-            while lineStart < glyphEnd {
-                var lineGlyphRange = NSRange(location: NSNotFound, length: 0)
-                let lineRect = layoutManager.lineFragmentRect(forGlyphAt: lineStart, effectiveRange: &lineGlyphRange, withoutAdditionalLayout: true)
-                guard lineGlyphRange.location != NSNotFound else { break }
-
-                let bandY = lineRect.origin.y + insetY
-                let bandHeight = lineRect.height
-                guard bandHeight > 0 else {
-                    lineStart = lineGlyphRange.location + lineGlyphRange.length
-                    continue
-                }
-
-                let bandRect = NSRect(x: insetX, y: bandY, width: containerWidth, height: bandHeight)
-                if bandRect.intersects(rect) {
-                    color.setFill()
-                    bandRect.fill()
-                }
-                lineStart = lineGlyphRange.location + lineGlyphRange.length
-            }
-            charIndex = effectiveRange.location + effectiveRange.length
-        }
-
-        // Decorative accent bar for blockquote ranges. Paragraph style supplies the
-        // leading indent (16pt); we paint a rounded bar of ~3pt in that gutter.
-        let barWidth: CGFloat = 3
-        let barInset: CGFloat = 4
-        charIndex = 0
-        while charIndex < length {
-            var effectiveRange = NSRange(location: NSNotFound, length: 0)
-            guard let color = storage.attribute(.blockquoteLeftBorder, at: charIndex, effectiveRange: &effectiveRange) as? NSColor,
-                  effectiveRange.location != NSNotFound else {
-                charIndex = effectiveRange.location != NSNotFound ? effectiveRange.location + effectiveRange.length : charIndex + 1
-                continue
-            }
-
-            let glyphRange = layoutManager.glyphRange(forCharacterRange: effectiveRange, actualCharacterRange: nil)
-            var lineStart = glyphRange.location
-            let glyphEnd = glyphRange.location + glyphRange.length
-
-            while lineStart < glyphEnd {
-                var lineGlyphRange = NSRange(location: NSNotFound, length: 0)
-                let lineRect = layoutManager.lineFragmentRect(forGlyphAt: lineStart, effectiveRange: &lineGlyphRange, withoutAdditionalLayout: true)
-                guard lineGlyphRange.location != NSNotFound else { break }
-
-                let bandHeight = lineRect.height
-                guard bandHeight > 0 else {
-                    lineStart = lineGlyphRange.location + lineGlyphRange.length
-                    continue
-                }
-
-                let barRect = NSRect(
-                    x: insetX + barInset,
-                    y: lineRect.origin.y + insetY,
-                    width: barWidth,
-                    height: bandHeight
-                )
-                if barRect.intersects(rect) {
-                    color.withAlphaComponent(0.75).setFill()
-                    let path = NSBezierPath(roundedRect: barRect, xRadius: barWidth / 2, yRadius: barWidth / 2)
-                    path.fill()
-                }
-                lineStart = lineGlyphRange.location + lineGlyphRange.length
-            }
-            charIndex = effectiveRange.location + effectiveRange.length
-        }
-    }
-
-    override func mouseDown(with event: NSEvent) {
-        if activatePaneOnReadOnlyInteraction(isEditable: isEditable, onActivatePane: onActivatePane) {
-            return
-        }
-        let point = convert(event.locationInWindow, from: nil)
-
-        if let hit = taskCheckboxTarget(at: point) {
-            _ = toggleTaskCheckbox(atCharacterIndex: hit.markerRange.location)
-            return
-        }
-
-        // Check if clicking on an image markdown
-        if let embedID = imageEmbedTarget(at: point) {
-            onSelectEmbed?(embedID)
-            return
-        }
-
-        if let target = wikilinkTarget(at: point) {
-            if wikilinkMarkdownIsHidden(at: point) {
-                let openInNewTab = event.modifierFlags.contains(.command)
-                _ = handleLinkClick(target, openInNewTab: openInNewTab)
-                return
-            }
-            // Markdown is exposed (always-on markdown mode, or the caret's active
-            // block under hide-while-typing): fall through so the click places the
-            // caret for editing instead of navigating.
-        }
-
-        // Check if clicking on a tag
-        if let tag = tagTarget(at: point) {
-            let openInNewTab = event.modifierFlags.contains(.command)
-            _ = handleTagClick(tag, openInNewTab: openInNewTab)
-            return
-        }
-        super.mouseDown(with: event)
-    }
-
-    private var trackingArea: NSTrackingArea?
-
-    override func updateTrackingAreas() {
-        super.updateTrackingAreas()
-
-        if let oldTrackingArea = trackingArea {
-            removeTrackingArea(oldTrackingArea)
-        }
-
-        let newTrackingArea = NSTrackingArea(
-            rect: bounds,
-            options: [.mouseMoved, .activeAlways, .inVisibleRect],
-            owner: self,
-            userInfo: nil
-        )
-        addTrackingArea(newTrackingArea)
-        trackingArea = newTrackingArea
-    }
-
-    override func mouseMoved(with event: NSEvent) {
-        let point = convert(event.locationInWindow, from: nil)
-
-        // Check if hovering over an interactive element
-        if taskCheckboxTarget(at: point) != nil ||
-           imageEmbedTarget(at: point) != nil ||
-           wikilinkTarget(at: point) != nil ||
-           tagTarget(at: point) != nil ||
-           urlTarget(at: point) != nil {
-            NSCursor.pointingHand.set()
-        } else {
-            NSCursor.arrow.set()
-        }
-    }
-
-    override func mouseExited(with event: NSEvent) {
-        NSCursor.arrow.set()
-    }
-
-    func imageEmbedTarget(at viewPoint: NSPoint) -> String? {
-        guard let layout = layoutManager, let container = textContainer else { return nil }
-
-        let containerPoint = NSPoint(
-            x: viewPoint.x - textContainerOrigin.x,
-            y: viewPoint.y - textContainerOrigin.y
-        )
-        var fraction: CGFloat = 0
-        let charIndex = layout.characterIndex(
-            for: containerPoint,
-            in: container,
-            fractionOfDistanceBetweenInsertionPoints: &fraction
-        )
-        guard charIndex < (string as NSString).length else { return nil }
-
-        let glyphIndex = layout.glyphIndexForCharacter(at: charIndex)
-        let glyphRect = layout.boundingRect(forGlyphRange: NSRange(location: glyphIndex, length: 1), in: container)
-        guard glyphRect.contains(containerPoint) else { return nil }
-
-        // Check if this character is part of an image markdown
-        let nsText = string as NSString
-        let textRange = NSRange(location: 0, length: nsText.length)
-
-        guard let regex = Self.inlineImageRegex else { return nil }
-        let matches = regex.matches(in: string, range: textRange)
-
-        for match in matches {
-            let matchRange = match.range(at: 0)
-            if NSLocationInRange(charIndex, matchRange) {
-                let source = nsText.substring(with: match.range(at: 2)).trimmingCharacters(in: .whitespacesAndNewlines)
-                return "\(matchRange.location)-\(source)"
-            }
-        }
-
-        return nil
-    }
-
-    func wikilinkTarget(at viewPoint: NSPoint) -> String? {
-        guard let layout = layoutManager, let container = textContainer else { return nil }
-
-        let containerPoint = NSPoint(
-            x: viewPoint.x - textContainerOrigin.x,
-            y: viewPoint.y - textContainerOrigin.y
-        )
-        var fraction: CGFloat = 0
-        let charIndex = layout.characterIndex(
-            for: containerPoint,
-            in: container,
-            fractionOfDistanceBetweenInsertionPoints: &fraction
-        )
-        guard charIndex < (string as NSString).length else { return nil }
-
-        let glyphIndex = layout.glyphIndexForCharacter(at: charIndex)
-        let glyphRect = layout.boundingRect(forGlyphRange: NSRange(location: glyphIndex, length: 1), in: container)
-        guard glyphRect.contains(containerPoint) else { return nil }
-
-        return textStorage?.attribute(.wikilinkTarget, at: charIndex, effectiveRange: nil) as? String
-    }
-
-    /// Whether the `[[...]]` markdown for the link at `viewPoint` is currently hidden,
-    /// which is the only case a WikiLink click should navigate. The markdown is hidden
-    /// only when hide-while-typing is on AND the click lands outside the caret's block
-    /// (the caret's block is revealed/exposed). In always-on markdown mode the syntax is
-    /// visible everywhere, so this returns false and clicks never navigate.
-    func wikilinkMarkdownIsHidden(at viewPoint: NSPoint) -> Bool {
-        guard settings?.hideMarkdownWhileEditing == true else { return false }
-        guard let layout = layoutManager, let container = textContainer else { return true }
-
-        let containerPoint = NSPoint(
-            x: viewPoint.x - textContainerOrigin.x,
-            y: viewPoint.y - textContainerOrigin.y
-        )
-        var fraction: CGFloat = 0
-        let charIndex = layout.characterIndex(
-            for: containerPoint,
-            in: container,
-            fractionOfDistanceBetweenInsertionPoints: &fraction
-        )
-
-        // The caret's block has its markdown revealed; every other block is collapsed.
-        // No active block (read-only / unfocused pane) ⇒ all links collapsed ⇒ clickable.
-        let reveal = MarkdownPreviewBlockReveal.make(
-            from: textStorage?.string ?? string,
-            cursorLocation: selectedRange().location,
-            isEditable: isEditable
-        )
-        guard let blockRange = reveal.blockRange else { return true }
-        return !NSLocationInRange(charIndex, blockRange)
-    }
-
-    func tagTarget(at viewPoint: NSPoint) -> String? {
-        guard let layout = layoutManager, let container = textContainer else { return nil }
-
-        let containerPoint = NSPoint(
-            x: viewPoint.x - textContainerOrigin.x,
-            y: viewPoint.y - textContainerOrigin.y
-        )
-        var fraction: CGFloat = 0
-        let charIndex = layout.characterIndex(
-            for: containerPoint,
-            in: container,
-            fractionOfDistanceBetweenInsertionPoints: &fraction
-        )
-        guard charIndex < (string as NSString).length else { return nil }
-
-        let glyphIndex = layout.glyphIndexForCharacter(at: charIndex)
-        let glyphRect = layout.boundingRect(forGlyphRange: NSRange(location: glyphIndex, length: 1), in: container)
-        guard glyphRect.contains(containerPoint) else { return nil }
-
-        return textStorage?.attribute(.tagTarget, at: charIndex, effectiveRange: nil) as? String
-    }
-
-    func urlTarget(at viewPoint: NSPoint) -> URL? {
-        guard let layout = layoutManager, let container = textContainer else { return nil }
-
-        let containerPoint = NSPoint(
-            x: viewPoint.x - textContainerOrigin.x,
-            y: viewPoint.y - textContainerOrigin.y
-        )
-        var fraction: CGFloat = 0
-        let charIndex = layout.characterIndex(
-            for: containerPoint,
-            in: container,
-            fractionOfDistanceBetweenInsertionPoints: &fraction
-        )
-        guard charIndex < (string as NSString).length else { return nil }
-
-        let glyphIndex = layout.glyphIndexForCharacter(at: charIndex)
-        let glyphRect = layout.boundingRect(forGlyphRange: NSRange(location: glyphIndex, length: 1), in: container)
-        guard glyphRect.contains(containerPoint) else { return nil }
-
-        return textStorage?.attribute(.link, at: charIndex, effectiveRange: nil) as? URL
-    }
-
-    // MARK: - Focus support
-
-    private var focusObserver: Any?
-
-    func installFocusObserver() {
-        guard focusObserver == nil else { return }
-        focusObserver = NotificationCenter.default.addObserver(
-            forName: .focusEditor,
-            object: nil,
-            queue: .main
-        ) { [weak self] _ in
-            guard let self, self.isEditable else { return }
-            preserveScrollOffset(for: self) {
-                self.window?.makeFirstResponder(self)
-            }
-        }
-    }
-
-    private var saveCursorObserver: Any?
-
-    func installSaveCursorObserver(appState: AppState) {
-        guard saveCursorObserver == nil else { return }
-        saveCursorObserver = NotificationCenter.default.addObserver(
-            forName: .saveCursorPosition,
-            object: nil,
-            queue: .main
-        ) { [weak self, weak appState] _ in
-            guard let self, self.isEditable, let appState else { return }
-            appState.pendingCursorRange = self.selectedRange()
-            appState.pendingScrollOffsetY = self.enclosingScrollView?.contentView.bounds.origin.y ?? 0
-        }
-    }
-
-    // MARK: - CMD-K observer
-
-    private var commandKObserver: Any?
-
-    func installCommandKObserver() {
-        guard commandKObserver == nil else { return }
-        commandKObserver = NotificationCenter.default.addObserver(
-            forName: .commandKPressed,
-            object: nil,
-            queue: .main
-        ) { [weak self] _ in
-            guard let self, self.isEditable else {
-                self?.onCommandPaletteFallback?()
-                return
-            }
-            let sel = self.selectedRange()
-            if sel.length > 0,
-               let selectedText = (self.string as NSString?)?.substring(with: sel),
-               !selectedText.isEmpty {
-                self.pendingWikilinkAlias = selectedText
-                self.pendingWikilinkSelectionRange = sel
-                self.onWikiLinkRequest?()
-            } else {
-                self.onCommandPaletteFallback?()
-            }
-        }
-    }
-
-    // MARK: - Search highlight support
-
-    private var searchObserver: Any?
-    private var searchClearObserver: Any?
-    private var replaceCurrentObserver: Any?
-    private var replaceAllObserver: Any?
-    private var lastSearchHighlightRanges: [NSRange] = []
-    private var lastSearchFocusIndex: Int = -1
-    /// The parsed block range whose markdown is currently revealed under the caret.
-    /// Used to skip re-revealing while the caret stays within one block.
-    private var lastRevealedBlockRange: NSRange?
-
-    /// Clears the revealed-block gate so the next revealCurrentBlockMarkdownAtCursor()
-    /// recomputes. Used after a full re-hide sweep that invalidated the visible reveal.
-    func invalidateRevealedBlock() {
-        lastRevealedBlockRange = nil
-    }
-
-    func installSearchObservers() {
-        guard searchObserver == nil else { return }
-        searchObserver = NotificationCenter.default.addObserver(
-            forName: .scrollToSearchMatch,
-            object: nil,
-            queue: .main
-        ) { [weak self] note in
-            guard let self,
-                  let query = note.userInfo?[SearchMatchKey.query] as? String,
-                  let focusIndex = note.userInfo?[SearchMatchKey.matchIndex] as? Int else { return }
-            self.applySearchHighlights(query: query, focusIndex: focusIndex)
-        }
-        searchClearObserver = NotificationCenter.default.addObserver(
-            forName: .clearSearchHighlights,
-            object: nil,
-            queue: .main
-        ) { [weak self] _ in
-            self?.clearSearchHighlights()
-        }
-        replaceCurrentObserver = NotificationCenter.default.addObserver(
-            forName: .replaceCurrentMatch,
-            object: nil,
-            queue: .main
-        ) { [weak self] note in
-            guard let self, self.participatesInGlobalSearch, self.isEditable,
-                  let query = note.userInfo?[SearchMatchKey.query] as? String,
-                  let focusIndex = note.userInfo?[SearchMatchKey.matchIndex] as? Int,
-                  let replacement = note.userInfo?[SearchMatchKey.replacement] as? String else { return }
-            let advanceAfter = (note.userInfo?[SearchMatchKey.advanceAfter] as? Bool) ?? false
-            self.replaceCurrentMatch(query: query, focusIndex: focusIndex, replacement: replacement, advanceAfter: advanceAfter)
-        }
-        replaceAllObserver = NotificationCenter.default.addObserver(
-            forName: .replaceAllMatches,
-            object: nil,
-            queue: .main
-        ) { [weak self] note in
-            guard let self, self.participatesInGlobalSearch, self.isEditable,
-                  let query = note.userInfo?[SearchMatchKey.query] as? String,
-                  let replacement = note.userInfo?[SearchMatchKey.replacement] as? String else { return }
-            self.replaceAllMatches(query: query, replacement: replacement)
-        }
-    }
-
-    private func applySearchHighlights(query: String, focusIndex: Int) {
-        guard let storage = textStorage, !query.isEmpty else {
-            clearSearchHighlights()
-            return
-        }
-        let content = storage.string
-        let needle = query.lowercased()
-        var matches: [NSRange] = []
-        var searchStart = content.startIndex
-        while searchStart < content.endIndex,
-              let range = content.range(of: needle, options: .caseInsensitive, range: searchStart..<content.endIndex) {
-            matches.append(NSRange(range, in: content))
-            searchStart = range.upperBound
-            if matches.count > 2000 { break }
-        }
-
-        let dimHighlight = NSColor.yellow.withAlphaComponent(0.30)
-        let focusHighlight = NSColor.yellow
-        storage.beginEditing()
-        let storageLength = storage.length
-        for range in lastSearchHighlightRanges {
-            // Ranges may be stale relative to the current storage (e.g. after an
-            // external edit). Skip any that no longer fit so removeAttribute can't
-            // throw NSRangeException and abort the rest of the highlight update.
-            guard NSMaxRange(range) <= storageLength else { continue }
-            storage.removeAttribute(.backgroundColor, range: range)
-        }
-        for (i, range) in matches.enumerated() {
-            if i == focusIndex {
-                storage.addAttribute(.backgroundColor, value: focusHighlight, range: range)
-                storage.addAttribute(.foregroundColor, value: NSColor.black, range: range)
-            } else {
-                storage.addAttribute(.backgroundColor, value: dimHighlight, range: range)
-            }
-        }
-        storage.endEditing()
-        lastSearchHighlightRanges = matches
-        lastSearchFocusIndex = focusIndex
-
-        // Report match count back to SwiftUI
-        onMatchCountUpdate?(matches.count)
-
-        // Scroll focused match into view (don't select — selection rendering overwrites highlight attributes)
-        if matches.indices.contains(focusIndex) {
-            scrollRangeToVisible(matches[focusIndex])
-        }
-    }
-
-    private func clearSearchHighlights() {
-        guard let storage = textStorage else { return }
-        storage.beginEditing()
-        let storageLength = storage.length
-        for range in lastSearchHighlightRanges {
-            guard NSMaxRange(range) <= storageLength else { continue }
-            storage.removeAttribute(.backgroundColor, range: range)
-        }
-        storage.endEditing()
-        lastSearchHighlightRanges = []
-        lastSearchFocusIndex = -1
-        applyMarkdownStyling()
-    }
-
-    private func replaceCurrentMatch(query: String, focusIndex: Int, replacement: String, advanceAfter: Bool) {
-        guard !query.isEmpty,
-              lastSearchHighlightRanges.indices.contains(focusIndex) else { return }
-        let range = lastSearchHighlightRanges[focusIndex]
-        guard shouldChangeText(in: range, replacementString: replacement) else { return }
-        textStorage?.replaceCharacters(in: range, with: replacement)
-        didChangeText()
-
-        // Recompute matches against new text. Anchor on the position of the replacement
-        // so the next focused match is the one that was after the replaced range.
-        let newCaret = range.location + (replacement as NSString).length
-        let newFocus: Int
-        if advanceAfter {
-            newFocus = nextMatchIndex(forQuery: query, after: newCaret)
-        } else {
-            newFocus = nextMatchIndex(forQuery: query, after: range.location)
-        }
-        applySearchHighlights(query: query, focusIndex: newFocus)
-    }
-
-    private func replaceAllMatches(query: String, replacement: String) {
-        guard !query.isEmpty, let storage = textStorage else { return }
-        let content = storage.string
-        // Use Foundation's single-pass replace instead of collecting every match range.
-        // A note can contain millions of occurrences of a short query; materializing
-        // `[NSRange]` for each would spike memory and freeze the main thread.
-        let mutable = NSMutableString(string: content)
-        let initialSearchRange = NSRange(location: 0, length: (mutable as NSString).length)
-        let replacedCount = mutable.replaceOccurrences(
-            of: query,
-            with: replacement,
-            options: .caseInsensitive,
-            range: initialSearchRange
-        )
-        let resultString = mutable as String
-
-        guard replacedCount > 0 else {
-            applySearchHighlights(query: query, focusIndex: 0)
-            return
-        }
-
-        let fullRange = NSRange(location: 0, length: storage.length)
-        guard shouldChangeText(in: fullRange, replacementString: resultString) else { return }
-        // Highlight ranges are about to be invalidated by the full-document replace.
-        // Drop them now so the debounced restyle (which re-applies highlights via
-        // reapplySearchHighlights) can't read out-of-bounds NSRanges and crash.
-        lastSearchHighlightRanges = []
-        lastSearchFocusIndex = -1
-        storage.replaceCharacters(in: fullRange, with: resultString)
-        didChangeText()
-
-        applySearchHighlights(query: query, focusIndex: 0)
-    }
-
-    /// Returns the index of the first match whose range starts at or after `location`,
-    /// wrapping to 0 if none. Recomputes matches against the live text storage.
-    private func nextMatchIndex(forQuery query: String, after location: Int) -> Int {
-        guard let storage = textStorage else { return 0 }
-        let content = storage.string
-        var matches: [NSRange] = []
-        var searchStart = content.startIndex
-        while searchStart < content.endIndex,
-              let r = content.range(of: query, options: .caseInsensitive, range: searchStart..<content.endIndex) {
-            matches.append(NSRange(r, in: content))
-            searchStart = r.upperBound
-            if matches.count > 2000 { break }
-        }
-        if matches.isEmpty { return 0 }
-        if let idx = matches.firstIndex(where: { $0.location >= location }) {
-            return idx
-        }
-        return 0
-    }
-
-    private func reapplySearchHighlights() {
-        guard !lastSearchHighlightRanges.isEmpty, let storage = textStorage else { return }
-        let dimHighlight = NSColor.yellow.withAlphaComponent(0.30)
-        let focusHighlight = NSColor.yellow
-        storage.beginEditing()
-        let storageLength = storage.length
-        for (i, range) in lastSearchHighlightRanges.enumerated() {
-            guard NSMaxRange(range) <= storageLength else { continue }
-            if i == lastSearchFocusIndex {
-                storage.addAttribute(.backgroundColor, value: focusHighlight, range: range)
-                storage.addAttribute(.foregroundColor, value: NSColor.black, range: range)
-            } else {
-                storage.addAttribute(.backgroundColor, value: dimHighlight, range: range)
-            }
-        }
-        storage.endEditing()
-    }
-
-    override func setFrameSize(_ newSize: NSSize) {
-        super.setFrameSize(newSize)
-        DispatchQueue.main.async { [weak self] in
-            self?.refreshInlineImagePreviews()
-            self?.refreshCollapsibleToggles()
-            self?.refreshCodeBlockCopyButtons()
-            self?.refreshAISparkle()
-        }
-    }
-
-    // MARK: - Block indent / dedent
-
-    private static let indentString = "    " // 4 spaces
-
-    /// Tab with a multi-line selection → indent every selected line.
-    /// Tab with a cursor or single-line selection → insert a literal tab (default).
-    override func insertTab(_ sender: Any?) {
-        let sel = selectedRange()
-        let nsText = string as NSString
-
-        // Determine whether the selection spans more than one line.
-        let selText = sel.length > 0 ? nsText.substring(with: sel) : ""
-        let spansMultipleLines = selText.contains("\n")
-
-        guard spansMultipleLines else {
-            super.insertTab(sender)
-            return
-        }
-
-        indentSelectedLines(dedent: false)
-    }
-
-    /// Shift-Tab: dedent every line touched by the selection.
-    /// Intercept via keyDown so we catch the Shift modifier.
-    private func indentSelectedLines(dedent: Bool) {
-        guard let storage = textStorage else { return }
-        let nsText = string as NSString
-        let sel = selectedRange()
-
-        // Expand selection to cover full lines.
-        let linesRange = nsText.lineRange(for: sel)
-
-        let linesText = nsText.substring(with: linesRange)
-        var lines = linesText.components(separatedBy: "\n")
-
-        // The last component after the trailing newline is always an empty
-        // string artifact — keep it so we don't drop the terminating newline.
-        let indent = Self.indentString
-
-        var newLines: [String] = []
-        for (i, line) in lines.enumerated() {
-            // Don't modify the empty artifact at the end.
-            if i == lines.count - 1 && line.isEmpty {
-                newLines.append(line)
-                continue
-            }
-            if dedent {
-                if line.hasPrefix(indent) {
-                    newLines.append(String(line.dropFirst(indent.count)))
-                } else if line.hasPrefix("\t") {
-                    newLines.append(String(line.dropFirst(1)))
-                } else {
-                    newLines.append(line) // nothing to dedent
-                }
-            } else {
-                newLines.append(indent + line)
-            }
-        }
-
-        let newText = newLines.joined(separator: "\n")
-        if shouldChangeText(in: linesRange, replacementString: newText) {
-            storage.beginEditing()
-            storage.replaceCharacters(in: linesRange, with: newText)
-            storage.endEditing()
-            didChangeText()
-
-            // Restore a selection that covers the same lines.
-            let newLinesRange = NSRange(location: linesRange.location, length: (newText as NSString).length)
-            setSelectedRange(newLinesRange)
-        }
-    }
-
-    override func insertNewline(_ sender: Any?) {
-        // Preserve the leading whitespace of the current line on the new line,
-        // and continue bullet lists (- or *) automatically.
-        let cursor = selectedRange().location
-        let nsText = string as NSString
-        guard cursor != NSNotFound else { super.insertNewline(sender); return }
-
-        // Find the start of the current line.
-        let lineRange = nsText.lineRange(for: NSRange(location: cursor, length: 0))
-        let lineText = nsText.substring(with: lineRange)
-            .replacingOccurrences(of: "\n", with: "")
-            .replacingOccurrences(of: "\r", with: "")
-
-        // Measure leading whitespace.
-        var indentEnd = lineText.startIndex
-        for ch in lineText {
-            if ch == " " || ch == "\t" { indentEnd = lineText.index(after: indentEnd) }
-            else { break }
-        }
-        let indent = String(lineText[lineText.startIndex..<indentEnd])
-        let afterIndent = String(lineText[indentEnd...])
-
-        // Detect bullet marker: "- " or "* " (unordered list items).
-        // Also handle "- [ ] " and "- [x] " task list items.
-        let bulletMarkers = ["- [ ] ", "- [x] ", "- [X] ", "* [ ] ", "* [x] ", "* [X] ", "- ", "* "]
-        var detectedMarker: String? = nil
-        for marker in bulletMarkers {
-            if afterIndent.hasPrefix(marker) {
-                detectedMarker = marker
-                break
-            }
-        }
-
-        // Detect ordered list: "1. ", "2. ", etc.
-        if detectedMarker == nil {
-            let orderedRegex = try? NSRegularExpression(pattern: #"^(\d+)\. "#)
-            let afterIndentNS = afterIndent as NSString
-            if let match = orderedRegex?.firstMatch(in: afterIndent, range: NSRange(location: 0, length: afterIndentNS.length)) {
-                let numberRange = match.range(at: 1)
-                let currentNumber = Int(afterIndentNS.substring(with: numberRange)) ?? 1
-                let markerLength = match.range(at: 0).length
-                let itemContent = String(afterIndent.dropFirst(markerLength))
-
-                if itemContent.trimmingCharacters(in: .whitespaces).isEmpty {
-                    // Empty ordered item — remove it and break out of the list.
-                    let deletionRange = NSRange(location: lineRange.location, length: cursor - lineRange.location)
-                    if shouldChangeText(in: deletionRange, replacementString: "") {
-                        replaceCharacters(in: deletionRange, with: "")
-                        didChangeText()
-                    }
-                    super.insertNewline(sender)
-                    return
-                }
-
-                super.insertNewline(sender)
-                insertText(indent + "\(currentNumber + 1). ", replacementRange: selectedRange())
-                return
-            }
-        }
-
-        guard let marker = detectedMarker else {
-            // No bullet — just continue with indent as before.
-            super.insertNewline(sender)
-            if !indent.isEmpty { insertText(indent, replacementRange: selectedRange()) }
-            return
-        }
-
-        let bulletContent = String(afterIndent.dropFirst(marker.count))
-
-        // If the bullet line is empty (user pressed enter on a blank bullet),
-        // remove the bullet and insert a plain newline instead.
-        if bulletContent.trimmingCharacters(in: .whitespaces).isEmpty {
-            // Delete back to the start of the bullet line and insert a bare newline.
-            let deletionRange = NSRange(location: lineRange.location, length: cursor - lineRange.location)
-            if shouldChangeText(in: deletionRange, replacementString: "") {
-                replaceCharacters(in: deletionRange, with: "")
-                didChangeText()
-            }
-            super.insertNewline(sender)
-            return
-        }
-
-        // Otherwise continue the list: new line with same indent + same marker.
-        // For task items, always start unchecked.
-        let continuationMarker: String
-        if marker.hasPrefix("- [") || marker.hasPrefix("* [") {
-            let bulletChar = marker.hasPrefix("-") ? "-" : "*"
-            continuationMarker = "\(bulletChar) [ ] "
-        } else {
-            continuationMarker = marker
-        }
-
-        super.insertNewline(sender)
-        insertText(indent + continuationMarker, replacementRange: selectedRange())
-    }
-
-    override func keyDown(with event: NSEvent) {
-        // ⌥J opens the inline AI bar at the cursor/selection (same as clicking the ✨).
-        // Works even when the ✨ is hidden via Settings.
-        let flags = event.modifierFlags.intersection(.deviceIndependentFlagsMask)
-        if flags == .option, event.charactersIgnoringModifiers?.lowercased() == "j" {
-            aiSparkleTapped()
-            return
-        }
-        if let popover = completionPopover, popover.isShown {
-            switch event.keyCode {
-            case KeyCode.downArrow: completionVC?.moveSelection(by: 1);    return
-            case KeyCode.upArrow: completionVC?.moveSelection(by: -1);     return
-            case KeyCode.returnKey, KeyCode.numpadEnter: completionVC?.selectCurrentItem(); return
-            case KeyCode.escape: dismissCompletion();                      return
-            default: break
-            }
-        }
-        // Shift-Tab on a multi-line selection → dedent.
-        if event.keyCode == KeyCode.tab, event.modifierFlags.contains(.shift) {
-            let sel = selectedRange()
-            let selText = sel.length > 0 ? (string as NSString).substring(with: sel) : ""
-            if selText.contains("\n") {
-                indentSelectedLines(dedent: true)
-                return
-            }
-        }
-        super.keyDown(with: event)
-    }
-
-    override func performKeyEquivalent(with event: NSEvent) -> Bool {
-        let flags = event.modifierFlags.intersection(.deviceIndependentFlagsMask)
-        if flags == .command, event.charactersIgnoringModifiers?.lowercased() == "v" {
-            // Only intercept paste when this text view (or one of its descendants) is the first responder.
-            // If focus is on a terminal or browser pane, let the event pass through naturally.
-            let responder = window?.firstResponder
-            let isSelfFocused: Bool
-            if let view = responder as? NSView {
-                isSelfFocused = view === self || view.isDescendant(of: self)
-            } else {
-                isSelfFocused = responder === self
-            }
-            guard isSelfFocused else { return false }
-            paste(self)
-            return true
-        }
-        // CMD-K with a non-empty selection: open the wikilink picker and use the
-        // selected text as the alias, so the result is [[picked-note|selected text]].
-        if flags == .command, event.charactersIgnoringModifiers?.lowercased() == "k" {
-            let sel = selectedRange()
-            if sel.length > 0, let selectedText = (string as NSString?)?.substring(with: sel), !selectedText.isEmpty {
-                pendingWikilinkAlias = selectedText
-                pendingWikilinkSelectionRange = sel
-                onWikiLinkRequest?()
-                return true
-            }
-        }
-        // CMD-Shift-K: delete the active line.
-        if flags == [.command, .shift], event.charactersIgnoringModifiers?.lowercased() == "k" {
-            let cursor = selectedRange().location
-            guard cursor != NSNotFound else { return false }
-            let nsText = string as NSString
-            let lineRange = nsText.lineRange(for: NSRange(location: cursor, length: 0))
-            var deletionRange = lineRange
-            var cursorPos = lineRange.location
-            if lineRange.length > 0 {
-                let lastChar = nsText.substring(with: NSRange(location: lineRange.location + lineRange.length - 1, length: 1))
-                if lastChar != "\n" && lastChar != "\r\n" && lineRange.location > 0 {
-                    deletionRange = NSRange(location: lineRange.location - 1, length: lineRange.length + 1)
-                    cursorPos = lineRange.location - 1
-                }
-            }
-            if shouldChangeText(in: deletionRange, replacementString: "") {
-                replaceCharacters(in: deletionRange, with: "")
-                didChangeText()
-                setSelectedRange(NSRange(location: cursorPos, length: 0))
-            }
-            return true
-        }
-        return super.performKeyEquivalent(with: event)
-    }
-
-    func prettifyTableIfNeeded() {
-        guard isEditable, !isPrettifyingTable else { return }
-        let cursor = selectedRange().location
-        guard cursor != NSNotFound else { return }
-
-        let source = string
-        let parser = MarkdownDocumentParser()
-        let document = parser.parse(source)
-
-        // Only prettify when cursor is OUTSIDE all tables
-        let cursorInTable = document.blocks.contains {
-            guard case .table = $0.kind else { return false }
-            return NSLocationInRange(cursor, $0.range)
-        }
-        guard !cursorInTable else { return }
-
-        let nsSource = source as NSString
-        isPrettifyingTable = true
-        defer { isPrettifyingTable = false }
-
-        // Prettify all tables in the document
-        let tableBlocks = document.blocks.filter { if case .table = $0.kind { return true }; return false }
-        // Process in reverse so earlier ranges stay valid after replacements
-        for tableBlock in tableBlocks.reversed() {
-            let tableText = nsSource.substring(with: tableBlock.range)
-            guard let result = MarkdownTablePrettifier.prettify(
-                tableText: tableText,
-                cursorOffsetInTable: 0
-            ) else { continue }
-            guard result.formatted != tableText else { continue }
-
-            if shouldChangeText(in: tableBlock.range, replacementString: result.formatted) {
-                replaceCharacters(in: tableBlock.range, with: result.formatted)
-                didChangeText()
-            }
-        }
-    }
-
-    func expandSlashCommandIfNeeded() {
-        let cursor = selectedRange().location
-        guard cursor != NSNotFound,
-              let context = slashCommandContext(in: string, cursor: cursor),
-              let command = SlashCommand(rawValue: context.query) else { return }
-
-        let output = resolveSlashCommandOutput(
-            command,
-            context: SlashCommandResolverContext(
-                now: slashCommandNowProvider(),
-                currentFileURL: currentFileURL,
-                locale: Locale(identifier: "en_US_POSIX"),
-                timeZone: slashCommandTimeZone
-            )
-        )
-        guard shouldChangeText(in: context.range, replacementString: output) else { return }
-        replaceCharacters(in: context.range, with: output)
-        didChangeText()
-        setSelectedRange(NSRange(location: context.range.location + (output as NSString).length, length: 0))
-    }
-
-    func checkForLinkTrigger(plainText: String? = nil, cursor cursorOverride: Int? = nil) {
-        let text = plainText ?? string
-        let nsText = text as NSString
-        var cursor = cursorOverride ?? selectedRange().location
-        guard cursor != NSNotFound else { dismissCompletion(); return }
-        cursor = min(max(0, cursor), nsText.length)
-        guard cursor > 0 else { dismissCompletion(); return }
-
-        // Some NSTextView edit notifications report cursor after a trailing paragraph newline.
-        while cursor > 0 {
-            let ch = nsText.substring(with: NSRange(location: cursor - 1, length: 1))
-            if ch.rangeOfCharacter(from: .newlines) != nil { cursor -= 1 } else { break }
-        }
-
-        let startOffset = max(0, cursor - 400)
-        let searchRange = NSRange(location: startOffset, length: cursor - startOffset)
-        let sub = nsText.substring(with: searchRange) as NSString
-        let bracketRange = sub.range(of: "[[", options: .backwards)
-        if bracketRange.location != NSNotFound {
-            let absStart = startOffset + bracketRange.location
-            let tokenRange = NSRange(location: absStart, length: cursor - absStart)
-            let token = nsText.substring(with: tokenRange)
-            guard token.hasPrefix("[[") else { dismissCompletion(); return }
-            let query = String(token.dropFirst(2))
-                .trimmingCharacters(in: .newlines)
-                .trimmingCharacters(in: .whitespaces)
-            debugLog("query='\(query)' allFiles=\(allFiles.count)")
-            // Limit completion to the actively typed token only.
-            if !query.contains("]]") && query.count <= 120 {
-                linkTypingRange = tokenRange
-                // Don't re-open the picker if the user ESC'd it for this [[ token.
-                if wikilinkPickerSuppressed { return }
-                // Use command palette for wiki link picker instead of completion popover
-                onWikiLinkRequest?()
-                return
-            }
-        }
-        dismissCompletion()
-    }
-
-    func dismissCompletion() {
-        completionPopover?.close()
-        completionPopover = nil
-        completionVC = nil
-        linkTypingRange = nil
-        wikilinkPickerSuppressed = false
-        if let m = eventMonitor { NSEvent.removeMonitor(m); eventMonitor = nil }
-    }
-
-    func insertLink(_ url: URL) {
-        let name = url.deletingPathExtension().lastPathComponent
-        let alias = pendingWikilinkAlias
-        let aliasRange = pendingWikilinkSelectionRange
-        clearPendingWikilinkInsertion()
-
-        if linkTypingRange == nil, let alias, !alias.isEmpty {
-            // CMD-K with selected text: replace the selection with [[name|alias]].
-            let currentLength = (string as NSString).length
-            let selRange = aliasRange ?? selectedRange()
-            guard selRange.location != NSNotFound,
-                  selRange.location + selRange.length <= currentLength else { return }
-            let linkText = "[[\(name)|\(alias)]]"
-            if shouldChangeText(in: selRange, replacementString: linkText) {
-                replaceCharacters(in: selRange, with: linkText)
-                didChangeText()
-                let afterLink = selRange.location + (linkText as NSString).length
-                DispatchQueue.main.async { [weak self] in
-                    guard let self else { return }
-                    self.window?.makeFirstResponder(self)
-                    self.setSelectedRange(NSRange(location: afterLink, length: 0))
-                }
-            }
-            return
-        }
-
-        // Normal [[...]] typing flow.
-        guard let range = linkTypingRange else { return }
-        guard range.location >= 0, range.location + range.length <= (string as NSString).length else {
-            dismissCompletion()
-            return
-        }
-        let typed = (string as NSString).substring(with: range)
-        guard typed.hasPrefix("[["), !typed.contains("\n"), range.length <= 120 else {
-            dismissCompletion()
-            return
-        }
-        let linkText = "[[\(name)]]"
-        if shouldChangeText(in: range, replacementString: linkText) {
-            replaceCharacters(in: range, with: linkText)
-            didChangeText()
-            // Restore focus and place cursor after ]] once the palette has dismissed.
-            let afterLink = range.location + (linkText as NSString).length
-            DispatchQueue.main.async { [weak self] in
-                guard let self else { return }
-                self.window?.makeFirstResponder(self)
-                self.setSelectedRange(NSRange(location: afterLink, length: 0))
-            }
-        }
-        dismissCompletion()
-    }
-
-    func handleLinkClick(_ link: Any, openInNewTab: Bool) -> Bool {
-        if let url = link as? URL {
-            // Use injected callback if available, otherwise fall back to NSWorkspace
-            if let onOpenExternalURL = onOpenExternalURL {
-                onOpenExternalURL(url)
-            } else {
-                NSWorkspace.shared.open(url)
-            }
-            return true
-        }
-
-        guard let inner = link as? String else { return false }
-        // Strip alias and heading for resolution
-        let name = inner.components(separatedBy: "|").first
-            .flatMap { $0.components(separatedBy: "#").first }
-            .map { $0.trimmingCharacters(in: .whitespaces) } ?? inner
-
-        if let match = allFiles.first(where: { $0.deletingPathExtension().lastPathComponent.lowercased() == name.lowercased() }) {
-            onOpenFile?(match, openInNewTab)
-            return true
-        }
-
-        // Unresolved — create a new note with this name in the same folder as the current file.
-        onCreateNote?(name, currentFileURL?.deletingLastPathComponent())
-        return true
-    }
-
-    func handleTagClick(_ tag: String, openInNewTab: Bool) -> Bool {
-        guard !tag.isEmpty else { return false }
-        onOpenTag?(tag, openInNewTab)
-        return true
-    }
-
-    func refreshInlineImagePreviews() {
-        // Inline image previews disabled - images now only show in sidebar
-        // This function is kept for compatibility but does nothing
-    }
-
-    // MARK: - Embedded Notes
-
-    func inlineEmbedMatches() -> [InlineEmbedMatch] {
-        guard let regex = Self.embedRegex else { return [] }
-        let nsText = string as NSString
-        let range = NSRange(location: 0, length: nsText.length)
-
-        return regex.matches(in: string, range: range).compactMap { match in
-            guard match.numberOfRanges > 1 else { return nil }
-            let raw = nsText.substring(with: match.range(at: 1))
-            // Extract note name (before any pipe alias or heading anchor)
-            let noteName = raw
-                .components(separatedBy: "|").first?
-                .components(separatedBy: "#").first?
-                .trimmingCharacters(in: .whitespacesAndNewlines) ?? ""
-            guard !noteName.isEmpty else { return nil }
-
-            let fullRange = match.range(at: 0)
-            let paragraphRange = nsText.paragraphRange(for: fullRange)
-            let id = "\(fullRange.location)-\(noteName)"
-
-            // Find the note file
-            let normalizedName = noteName.lowercased()
-            let noteURL = allFiles.first { url in
-                url.deletingPathExtension().lastPathComponent.lowercased() == normalizedName
-            }
-
-            // Get content if note exists
-            var content: String?
-            if let noteURL = noteURL {
-                content = try? String(contentsOf: noteURL, encoding: .utf8)
-            }
-
-            return InlineEmbedMatch(
-                id: id,
-                range: fullRange,
-                paragraphRange: paragraphRange,
-                noteName: noteName,
-                content: content,
-                noteURL: noteURL
-            )
-        }
-    }
-
-    func inlineImageMatches() -> [InlineImageMatch] {
-        guard let regex = Self.inlineImageRegex else { return [] }
-        let nsText = string as NSString
-        let range = NSRange(location: 0, length: nsText.length)
-
-        return regex.matches(in: string, range: range).compactMap { match in
-            guard match.numberOfRanges > 2 else { return nil }
-            let caption = nsText.substring(with: match.range(at: 1))
-            let source = nsText.substring(with: match.range(at: 2)).trimmingCharacters(in: .whitespacesAndNewlines)
-            let fullRange = match.range(at: 0)
-            let paragraphRange = nsText.paragraphRange(for: fullRange)
-            return InlineImageMatch(
-                id: "\(fullRange.location)-\(source)",
-                range: fullRange,
-                paragraphRange: paragraphRange,
-                source: source,
-                caption: caption
-            )
-        }
-    }
-
-    func visibleInlineImageMatches() -> [InlineImageMatch] {
-        let matches = inlineImageMatches()
-        guard !matches.isEmpty else { return [] }
-
-        let fileURL = currentFileURL ?? AppConstants.unsavedFileURL
-        let sections = collapsibleParser.parse(string)
-        let collapsedRanges = sections.compactMap { section -> NSRange? in
-            guard section.contentRange.length > 0 else { return nil }
-            let sectionId = section.getIdentifier()
-            return collapsibleStateManager.isCollapsed(sectionId, in: fileURL) ? section.contentRange : nil
-        }
-
-        guard !collapsedRanges.isEmpty else { return matches }
-        return matches.filter { match in
-            !collapsedRanges.contains { NSIntersectionRange($0, match.range).length > 0 }
-        }
-    }
-
-    struct InlinePreviewAsset {
-        let image: NSImage
-        let imageDataType: NSPasteboard.PasteboardType
-        let preservesAnimation: Bool
-    }
-
-    func inlinePreviewAsset(fromFileURL url: URL, maxPixelSize: CGFloat) -> InlinePreviewAsset? {
-        guard let data = try? Data(contentsOf: url) else { return nil }
-        return inlinePreviewAsset(from: data, maxPixelSize: maxPixelSize)
-    }
-
-    func inlinePreviewAsset(from data: Data, maxPixelSize: CGFloat) -> InlinePreviewAsset? {
-        guard let source = CGImageSourceCreateWithData(data as CFData, nil) else { return nil }
-        let imageType = (CGImageSourceGetType(source) as String?) ?? "public.image"
-        let pasteboardType = NSPasteboard.PasteboardType(imageType)
-        let frameCount = CGImageSourceGetCount(source)
-        let preservesAnimation = (imageType == "com.compuserve.gif" || imageType == "public.gif") && frameCount > 1
-
-        if preservesAnimation, let image = NSImage(data: data) {
-            return InlinePreviewAsset(image: image, imageDataType: pasteboardType, preservesAnimation: true)
-        }
-
-        if let image = downsampledImage(from: source, maxPixelSize: maxPixelSize) ?? NSImage(data: data) {
-            return InlinePreviewAsset(image: image, imageDataType: pasteboardType, preservesAnimation: false)
-        }
-
-        return nil
-    }
-
-    private func downsampledImage(from source: CGImageSource, maxPixelSize: CGFloat) -> NSImage? {
-        let options: [CFString: Any] = [
-            kCGImageSourceCreateThumbnailFromImageAlways: true,
-            kCGImageSourceShouldCacheImmediately: true,
-            kCGImageSourceCreateThumbnailWithTransform: true,
-            kCGImageSourceThumbnailMaxPixelSize: max(400, Int(maxPixelSize)),
-        ]
-
-        guard let cgImage = CGImageSourceCreateThumbnailAtIndex(source, 0, options as CFDictionary) else {
-            return nil
-        }
-
-        let image = NSImage(cgImage: cgImage, size: .zero)
-        return image.size.width > 0 || image.size.height > 0 ? image : nil
-    }
-
-    // MARK: - Image paste handling
-
-    /// Handles paste events for images. Saves image to .images folder and inserts markdown.
-    override func paste(_ sender: Any?) {
-        // Image data takes priority: if the pasteboard carries actual image
-        // binary data, save it to .images/ and insert a local Markdown link.
-        // This must come before HTML so that copying an image from a browser
-        // (which puts both image data AND HTML on the pasteboard) correctly
-        // saves the image locally rather than emitting a remote URL.
-        if handlePaste(from: .general) {
-            return
-        }
-
-        // No image data — try converting HTML to Markdown.
-        if handleHTMLPaste(from: .general) {
-            return
-        }
-
-        super.paste(sender)
-    }
-
-    @discardableResult
-    func handlePaste(from pasteboard: NSPasteboard) -> Bool {
-        guard let asset = readPastedImageAsset(from: pasteboard) else {
-            return false
-        }
-        handleImagePaste(asset: asset)
-        return true
-    }
-
-    private struct PastedImageAsset {
-        let image: NSImage
-        let originalData: Data?
-        let fileExtension: String
-    }
-
-    private func readPastedImageAsset(from pasteboard: NSPasteboard) -> PastedImageAsset? {
-        let gifTypes: [NSPasteboard.PasteboardType] = [
-            NSPasteboard.PasteboardType(rawValue: "com.compuserve.gif"),
-            NSPasteboard.PasteboardType(rawValue: "public.gif"),
-            NSPasteboard.PasteboardType(rawValue: "GIF"),
-            NSPasteboard.PasteboardType(rawValue: "GIFf"),
-        ]
-
-        for type in gifTypes {
-            if let gifData = pasteboard.data(forType: type),
-               let image = NSImage(data: gifData) {
-                return PastedImageAsset(image: image, originalData: gifData, fileExtension: "gif")
-            }
-        }
-
-        if let fileURLs = pasteboard.readObjects(forClasses: [NSURL.self], options: [.urlReadingContentsConformToTypes: ["public.image"]]) as? [URL],
-           let firstURL = fileURLs.first,
-           let image = NSImage(contentsOf: firstURL) {
-            let ext = firstURL.pathExtension.lowercased()
-            if ext == "gif", let data = try? Data(contentsOf: firstURL) {
-                return PastedImageAsset(image: image, originalData: data, fileExtension: ext)
-            }
-            return PastedImageAsset(image: image, originalData: nil, fileExtension: "png")
-        }
-
-        if let urlData = pasteboard.data(forType: NSPasteboard.PasteboardType(rawValue: "public.file-url")),
-           let urlString = String(data: urlData, encoding: .utf8),
-           let url = URL(string: urlString),
-           let image = NSImage(contentsOf: url) {
-            let ext = url.pathExtension.lowercased()
-            if ext == "gif", let data = try? Data(contentsOf: url) {
-                return PastedImageAsset(image: image, originalData: data, fileExtension: ext)
-            }
-            return PastedImageAsset(image: image, originalData: nil, fileExtension: "png")
-        }
-
-        guard let image = readImage(from: pasteboard) else { return nil }
-        return PastedImageAsset(image: image, originalData: nil, fileExtension: "png")
-    }
-
-    /// Reads an image from the pasteboard using various methods
-    private func readImage(from pasteboard: NSPasteboard) -> NSImage? {
-        if let images = pasteboard.readObjects(forClasses: [NSImage.self]) as? [NSImage],
-           let firstImage = images.first {
-            return firstImage
-        }
-
-        if NSImage.canInit(with: pasteboard) {
-            if let image = NSImage(pasteboard: pasteboard) {
-                return image
-            }
-        }
-
-        let tiffTypes: [NSPasteboard.PasteboardType] = [
-            .tiff,
-            NSPasteboard.PasteboardType(rawValue: "public.tiff"),
-            NSPasteboard.PasteboardType(rawValue: "TIFF"),
-            NSPasteboard.PasteboardType(rawValue: "com.apple.tiff"),
-            NSPasteboard.PasteboardType(rawValue: "NeXT TIFF v4.0 pasteboard type"),
-        ]
-
-        for type in tiffTypes {
-            if let tiffData = pasteboard.data(forType: type) {
-                if let image = NSImage(data: tiffData) {
-                    return image
-                }
-            }
-        }
-
-        let pngTypes: [NSPasteboard.PasteboardType] = [
-            .png,
-            NSPasteboard.PasteboardType(rawValue: "public.png"),
-            NSPasteboard.PasteboardType(rawValue: "PNG"),
-            NSPasteboard.PasteboardType(rawValue: "PNGf"),
-            NSPasteboard.PasteboardType(rawValue: "Apple PNG pasteboard type"),
-        ]
-
-        for type in pngTypes {
-            if let pngData = pasteboard.data(forType: type) {
-                if let image = NSImage(data: pngData) {
-                    return image
-                }
-            }
-        }
-
-        let otherImageTypes: [NSPasteboard.PasteboardType] = [
-            NSPasteboard.PasteboardType(rawValue: "public.jpeg"),
-            NSPasteboard.PasteboardType(rawValue: "public.jpg"),
-            NSPasteboard.PasteboardType(rawValue: "JPEG"),
-            NSPasteboard.PasteboardType(rawValue: "JFIF"),
-            NSPasteboard.PasteboardType(rawValue: "public.image"),
-            NSPasteboard.PasteboardType(rawValue: "com.apple.pict"),
-            NSPasteboard.PasteboardType(rawValue: "GIF"),
-            NSPasteboard.PasteboardType(rawValue: "GIFf"),
-            NSPasteboard.PasteboardType(rawValue: "BMP"),
-            NSPasteboard.PasteboardType(rawValue: "BMPf"),
-        ]
-
-        for type in otherImageTypes {
-            if let data = pasteboard.data(forType: type),
-               let image = NSImage(data: data) {
-                return image
-            }
-        }
-
-        return nil
-    }
-
-    private func handleImagePaste(asset: PastedImageAsset) {
-        guard let currentFileURL = currentFileURL else {
-            // No current file, fall back to regular paste (but images can't be pasted without a file context)
-            return
-        }
-
-        let fileFolder = currentFileURL.deletingLastPathComponent()
-        let imagesFolder = fileFolder.appendingPathComponent(".images")
-
-        // Create .images folder if it doesn't exist
-        let fileManager = FileManager.default
-        if !fileManager.fileExists(atPath: imagesFolder.path) {
-            do {
-                try fileManager.createDirectory(at: imagesFolder, withIntermediateDirectories: true)
-            } catch {
-                debugLog("Failed to create .images folder: \(error)")
-                return
-            }
-        }
-
-        // Generate unique filename with timestamp and random component
-        let timestamp = Int(Date().timeIntervalSince1970)
-        let random = Int.random(in: 1000...9999)
-        let filename = "image_\(timestamp)_\(random).\(asset.fileExtension)"
-        let imagePath = imagesFolder.appendingPathComponent(filename)
-
-        let dataToWrite: Data
-        if let originalData = asset.originalData {
-            dataToWrite = originalData
-        } else {
-            guard let tiffData = asset.image.tiffRepresentation,
-                  let bitmap = NSBitmapImageRep(data: tiffData),
-                  let pngData = bitmap.representation(using: .png, properties: [:]) else {
-                debugLog("Failed to convert image to PNG")
-                return
-            }
-            dataToWrite = pngData
-        }
-
-        do {
-            try dataToWrite.write(to: imagePath)
-        } catch {
-            debugLog("Failed to save image: \(error)")
-            return
-        }
-
-        // Calculate relative path from current file to image
-        let relativePath = ".images/\(filename)"
-        let markdown = "![](\(relativePath))"
-
-        // Insert markdown at current cursor position
-        let currentRange = selectedRange()
-        if shouldChangeText(in: currentRange, replacementString: markdown) {
-            replaceCharacters(in: currentRange, with: markdown)
-            didChangeText()
-        }
-    }
-
-    // MARK: - HTML to Markdown Conversion
-
-    /// Returns true when the insertion point is inside (or on the opening line
-    /// of) a fenced code block, so HTML paste should be left as-is.
-    private func cursorIsInCodeBlock() -> Bool {
-        let cursor = selectedRange().location
-        guard cursor != NSNotFound else { return false }
-        let nsText = string as NSString
-        let len = nsText.length
-
-        // Pattern: opening ``` fence at the start of a line.
-        guard let fenceRegex = try? NSRegularExpression(
-            pattern: "^[ \\t]{0,3}```",
-            options: [.anchorsMatchLines]
-        ) else { return false }
-
-        let fenceMatches = fenceRegex.matches(in: string, options: [],
-                                              range: NSRange(location: 0, length: len))
-
-        // Walk fence matches in pairs; an unpaired opening also counts.
-        var i = 0
-        while i < fenceMatches.count {
-            let open = fenceMatches[i]
-            let openEnd = open.range.location + open.range.length
-
-            if i + 1 < fenceMatches.count {
-                let close = fenceMatches[i + 1]
-                let closeStart = close.range.location
-                // Cursor is inside a complete block.
-                if cursor >= open.range.location && cursor <= NSMaxRange(close.range) {
-                    return true
-                }
-                i += 2
-            } else {
-                // Unpaired opening fence — cursor anywhere from the ``` to end of doc.
-                if cursor >= open.range.location {
-                    return true
-                }
-                i += 1
-            }
-        }
-        return false
-    }
-
-    /// Converts HTML pasteboard content to Markdown and inserts it.
-    /// Checks dedicated HTML pasteboard types first, then falls back to
-    /// checking whether the plain-text payload looks like HTML (e.g. when
-    /// copying raw HTML source from a text editor or terminal).
-    @discardableResult
-    func handleHTMLPaste(from pasteboard: NSPasteboard) -> Bool {
-        // Never convert when the cursor is inside a fenced code block.
-        guard !cursorIsInCodeBlock() else { return false }
-
-        // 1. Try dedicated HTML pasteboard types (browser copies, rich-text apps).
-        let htmlTypes: [NSPasteboard.PasteboardType] = [
-            .html,
-            NSPasteboard.PasteboardType("public.html"),
-            NSPasteboard.PasteboardType("Apple HTML pasteboard type"),
-            NSPasteboard.PasteboardType("NSHTMLPboardType"),
-        ]
-
-        var htmlString: String? = nil
-        for type in htmlTypes {
-            if let str = pasteboard.string(forType: type) {
-                htmlString = str
-                break
-            }
-        }
-
-        // 2. Fallback: plain-text that looks like HTML (raw source pasted from
-        //    a terminal, VS Code, etc.). Require at least one structural tag so
-        //    we don't accidentally convert markdown or code that uses < >.
-        if htmlString == nil, let plain = pasteboard.string(forType: .string) {
-            if looksLikeHTML(plain) {
-                htmlString = plain
-            }
-        }
-
-        guard let html = htmlString, !html.isEmpty else {
-            return false
-        }
-
-        #if DEBUG
-        print("[HTML Paste] Source: \(html.prefix(300))")
-        #endif
-
-        let markdown = HTMLToMarkdownConverter.convert(html)
-
-        #if DEBUG
-        print("[HTML Paste] Result: \(markdown.prefix(300))")
-        #endif
-
-        let currentRange = selectedRange()
-        guard shouldChangeText(in: currentRange, replacementString: markdown) else {
-            return false
-        }
-        replaceCharacters(in: currentRange, with: markdown)
-        didChangeText()
-        return true
-    }
-
-    /// Returns true if the string contains at least one structural HTML tag
-    /// that makes it worth attempting conversion.
-    private func looksLikeHTML(_ text: String) -> Bool {
-        let structural = ["<ul", "<ol", "<li", "<p>", "<p ", "<h1", "<h2",
-                          "<h3", "<h4", "<h5", "<h6", "<table", "<div",
-                          "<blockquote", "<pre", "<code"]
-        let lower = text.lowercased()
-        return structural.contains { lower.contains($0) }
     }
 }
